% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{AnnenBatrute}
\alias{AnnenBatrute}
\title{AnnenBatrute}
\format{
\preformatted{Simple feature collection with 53 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -57681.01 ymin: 6734699 xmax: 1102790 ymax: 7941234
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 53 x 4
   målemetode nøyaktighet oppdateringsdato                                     geometry
 *      <int>       <int> <date>                                       <LINESTRING [m]>
 1         64       50000 2007-10-24                 (1012611 7922331, 1011518 7921259)
 2         64       50000 2004-06-01       (966769.2 7938989, 970232.1 7940345, 971955~
 3         64       50000 2004-06-01       (1012611 7922331, 1013836 7923732, 1014391 ~
 4         64       50000 2004-06-01       (1043402 7908353, 1044139 7910827, 1044706 ~
 5         64       50000 2004-06-01       (1097176 7888490, 1098501 7888274, 1099453 ~
 6         64       50000 2004-06-01       (1088500 7850137, 1084103 7847118, 1078877 ~
 7         64       50000 2004-06-01       (1075218 7823496, 1075284 7827537, 1075937 ~
 8         64       50000 2002-09-15       (1055103 7839692, 1057570 7838621, 1063510 ~
 9         64       50000 1997-06-30       (962099 7934023, 962907.3 7935718, 963479.5~
10         64       50000 2004-07-01       (962099 7934023, 961746 7936037, 961077.8 7~
# ... with 43 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Samferdsel_GML.gml}
}
\usage{
AnnenBatrute
}
\description{
AnnenBåtrute
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
