% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Arealbrukgrense}
\alias{Arealbrukgrense}
\title{Arealbrukgrense}
\format{
\preformatted{Simple feature collection with 14953 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -58988.99 ymin: 6454505 xmax: 1114893 ymax: 7921538
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 14,953 x 4
   målemetode nøyaktighet oppdateringsdato                                     geometry
 *      <int>       <int> <date>                                       <LINESTRING [m]>
 1         64       50000 2000-10-04       (905013 7868638, 905270 7868613, 905932 786~
 2         64       50000 2000-10-04       (905013 7868638, 904948 7868478, 904980 786~
 3         64       50000 2000-10-04       (1e+06 7869714, 1000424 7869137, 1000550 78~
 4         64       50000 2004-06-01       (986717 7867608, 986512 7867436, 986440.1 7~
 5         64       50000 2000-10-04       (991429 7866476, 991394 7866734, 991249.5 7~
 6         64       50000 2000-10-04       (1e+06 7866606, 999926 7866536, 1e+06 78663~
 7         64       50000 2000-10-04       (829228 7866857, 829164 7867184, 829032 786~
 8         64       50000 2000-10-04       (1036278 7865928, 1036892 7866288, 1037208 ~
 9         64       50000 2000-10-04       (1e+06 7866387, 1000139 7866108, 1e+06 7865~
10         64       50000 2000-10-04       (872246 7865227, 872476 7865294, 872683 786~
# ... with 14,943 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
Arealbrukgrense
}
\description{
Arealbrukgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
