% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Bane}
\alias{Bane}
\title{Bane}
\format{
\preformatted{Simple feature collection with 282 features and 4 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -42446 ymin: 6466305 xmax: 1076833 ymax: 7806691
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 282 x 5
   målemetode nøyaktighet medium oppdateringsdato                              geometry
 *      <int>       <int> <chr>  <date>                                <LINESTRING [m]>
 1         64       50000 U      2001-08-01       (4189.62 6513952, 3276.4 6513816, 19~
 2         64       50000 U      1993-01-01       (11050 6508417, 11105.87 6508365, 11~
 3         64       50000 T      1993-01-01       (14016 6504481, 14146 6504739, 14313~
 4         64       50000 T      2001-08-01       (13671.99 6505022, 13674.48 6504712,~
 5         64       50000 U      2001-08-01       (19693 6502988, 20588 6503192, 21096~
 6         64       50000 T      2001-08-01       (21096 6503039, 21270.83 6502365, 21~
 7         64       50000 T      1993-01-01       (29825 6499511, 30020.49 6499430, 30~
 8         64       50000 U      1993-01-01       (33285 6497952, 31823.56 6497983, 31~
 9         64       50000 U      1993-01-01       (81395 6468174, 82218 6467897, 83297~
10         64       50000 T      1993-01-01       (83297 6467829, 83906.8 6467658, 840~
# ... with 272 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Samferdsel_GML.gml}
}
\usage{
Bane
}
\description{
Bane
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
