% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Barmarksloype}
\alias{Barmarksloype}
\title{Barmarksloype}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 62 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 797893.2 ymin: 7650834 xmax: 1066724 ymax: 7873456
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 62 × 4
   målemetode nøyaktighet oppdateringsdato                                   senterlinje
 *      <int>       <int> <date>                                        <LINESTRING [m]>
 1         64       50000 2017-01-05       (1058259 7846052, 1058418 7846930, 1058449 7…
 2         64       50000 2017-01-05       (838375 7769214, 838796.8 7769149, 839237.7 …
 3         64       50000 2017-01-05       (908692 7867390, 908631.8 7867583, 908577.7 …
 4         64       50000 2017-01-05       (943577.1 7873259, 943734 7873331, 944078 78…
 5         64       50000 2017-01-05       (851211.9 7756127, 850794.7 7756375, 850444.…
 6         64       50000 2017-01-05       (1011624 7831045, 1011544 7831102, 1011457 7…
 7         64       50000 2017-01-05                (833431.6 7779698, 833577.6 7779595)
 8         64       50000 2017-01-05       (825584.6 7686748, 825069.2 7687199, 824849.…
 9         64       50000 2017-01-05       (1026513 7869564, 1026649 7869502, 1026718 7…
10         64       50000 2017-01-05       (1027698 7866585, 1027561 7866302, 1027391 7…
# ℹ 52 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Samferdsel_GML.gml}
}
\usage{
Barmarksloype
}
\description{
Barmarksløype
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
