% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Bilferjestrekning}
\alias{Bilferjestrekning}
\title{Bilferjestrekning}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 240 features and 6 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -74407.93 ymin: 6484917 xmax: 850374.3 ymax: 7904854
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 240 × 7
   målemetode nøyaktighet medium oppdateringsdato vegkategori vegnummer
 *      <int>       <int> <chr>  <date>           <chr>           <int>
 1         64       50000 T      2017-01-05       F                  17
 2         64       50000 T      2017-01-05       F                 839
 3         64       50000 T      2017-01-05       F                 839
 4         64       50000 T      2017-01-05       F                  17
 5         64       50000 T      2017-01-05       F                 469
 6         64       50000 T      2017-01-05       F                  81
 7         64       50000 T      2017-01-05       F                  81
 8         64       50000 T      2017-01-05       F                 552
 9         64       50000 T      2017-01-05       F                 769
10         64       50000 T      2017-01-05       F                 769
# ℹ 230 more rows
# ℹ 1 more variable: senterlinje <LINESTRING [m]>
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Samferdsel_GML.gml}
}
\usage{
Bilferjestrekning
}
\description{
Bilferjestrekning
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
