% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Bygning}
\alias{Bygning}
\title{Bygning}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 2701 features and 7 fields
Geometry type: POINT
Dimension:     XY
Bounding box:  xmin: -74711.91 ymin: 6454575 xmax: 1107651 ymax: 7930848
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 2,701 × 8
   målemetode nøyaktighet oppdateringsdato bygningskategori betjeningsgrad hytteeier
 *      <int>       <int> <date>                      <int> <chr>              <int>
 1         64       50000 2017-01-05                     10 <NA>                  NA
 2         64       50000 2017-01-05                     10 <NA>                  NA
 3         64       50000 2017-01-05                     10 <NA>                  NA
 4         64       50000 2017-01-05                     10 <NA>                  NA
 5         64       50000 2017-01-05                     10 <NA>                  NA
 6         64       50000 2017-01-05                     10 <NA>                  NA
 7         64       50000 2017-01-05                     10 <NA>                  NA
 8         64       50000 2017-01-05                     10 <NA>                  NA
 9         64       50000 2017-01-05                     10 <NA>                  NA
10         64       50000 2017-01-05                     10 <NA>                  NA
# ℹ 2,691 more rows
# ℹ 2 more variables: navn <chr>, posisjon <POINT [m]>
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000BygningerOgAnlegg_GML.gml}
}
\usage{
Bygning
}
\description{
Bygning
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
