% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Dam}
\alias{Dam}
\title{Dam}
\format{
\preformatted{Simple feature collection with 3 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 31646.77 ymin: 6542912 xmax: 42382.63 ymax: 6611421
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 3 x 4
  målemetode nøyaktighet oppdateringsdato                                      geometry
*      <int>       <int> <date>                                        <LINESTRING [m]>
1         64       50000 2001-08-01       (39006.2 6542912, 39683.26 6542993, 40373.84~
2         64       50000 2006-07-01       (31907.02 6611421, 31750.48 6611224, 31646.7~
3         64       50000 2001-08-01       (41143.16 6601383, 41182.02 6601384, 41183.6~
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000BygningerOgAnlegg_GML.gml}
}
\usage{
Dam
}
\description{
Dam
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
