% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Elv}
\alias{Elv}
\title{Elv}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 27 features and 0 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: 187462 ymin: 6566414 xmax: 1081802 ymax: 7873955
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 27 × 1
                                                                                  område
 *                                                                         <POLYGON [m]>
 1 ((221092 6658018, 220855 6658206, 220795 6658103, 220748 6658075, 220543 6657953, 22…
 2       ((3e+05 6665875, 300308 6665831, 300335 6666066, 3e+05 6666076, 3e+05 6665875))
 3 ((283322 6617295, 283181 6617402, 283158.6 6617373, 283099.7 6617297, 283017 6617191…
 4 ((273451 6577390, 273271 6577419, 273281.8 6577358, 273293 6577295, 273372.8 6576953…
 5 ((188469 6576217, 188151 6576449, 188014 6576560, 187587 6576684, 187462 6576641, 18…
 6 ((287438 6619581, 287409 6619803, 287000 6619731, 286510 6619679, 286265 6619615, 28…
 7 ((325715 7079814, 325819 7079754, 326081 7079195, 326288 7078513, 326292.9 7078497, …
 8 ((419112 7296642, 418890 7297273, 418280 7298306, 418280 7298306, 418242 7298485, 41…
 9 ((1062063 7733070, 1061917 7733327, 1061731 7733301, 1061965 7732954, 1062088 773302…
10 ((281383 6614488, 281304 6614558, 281150 6614369, 280822 6614308, 280804.2 6614294, …
# ℹ 17 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
Elv
}
\description{
Elv
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
