% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{ElvBekk}
\alias{ElvBekk}
\title{ElvBekk}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 9054 features and 4 fields (with 9027 geometries empty)
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: 187462 ymin: 6566414 xmax: 1081802 ymax: 7873955
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 9,054 × 5
   oppdateringsdato vannBredde målemetode nøyaktighet                             område
 * <date>                <int>      <int>       <int>                      <POLYGON [m]>
 1 2017-01-05                6         NA          NA ((221092 6658018, 220855 6658206,…
 2 2017-01-05                6         NA          NA ((3e+05 6665875, 300308 6665831, …
 3 2017-01-05                6         NA          NA ((283322 6617295, 283181 6617402,…
 4 2017-01-05                6         NA          NA ((273451 6577390, 273271 6577419,…
 5 2017-01-05                6         NA          NA ((188469 6576217, 188151 6576449,…
 6 2017-01-05                6         NA          NA ((287438 6619581, 287409 6619803,…
 7 2017-01-05                6         NA          NA ((325715 7079814, 325819 7079754,…
 8 2017-01-05                6         NA          NA ((419112 7296642, 418890 7297273,…
 9 2017-01-05                6         NA          NA ((1062063 7733070, 1061917 773332…
10 2017-01-05                6         NA          NA ((281383 6614488, 281304 6614558,…
# ℹ 9,044 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
ElvBekk
}
\description{
ElvBekk
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
