% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{ElvBekkKant}
\alias{ElvBekkKant}
\title{ElvBekkKant}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 386 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 187462 ymin: 6566414 xmax: 1081802 ymax: 7873955
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 386 × 4
   målemetode nøyaktighet oppdateringsdato                                        grense
 *      <int>       <int> <date>                                        <LINESTRING [m]>
 1         64       50000 2017-01-05       (991429 7866476, 991951 7867260, 992167 7867…
 2         64       50000 2017-01-05       (990838 7865195, 990862 7865333, 990970 7865…
 3         64       50000 2017-01-05                    (993185 7873440, 993516 7873955)
 4         64       50000 2017-01-05       (994613 7873673, 994441 7873143, 994343 7872…
 5         64       50000 2017-01-05       (992147 7869620, 992394 7870151, 992497 7870…
 6         64       50000 2017-01-05       (994592 7871079, 994683 7870068, 994623 7869…
 7         64       50000 2017-01-05       (990673 7860838, 990656 7861206, 991060 7861…
 8         64       50000 2017-01-05       (991721 7858486, 991540 7858660, 991493 7859…
 9         64       50000 2017-01-05       (995919 7840607, 996022 7840223, 996242 7839…
10         64       50000 2017-01-05       (993410 7867802, 993328 7867668, 993251 7867…
# ℹ 376 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
ElvBekkKant
}
\description{
ElvBekkKant
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
