% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{FiktivDelelinje}
\alias{FiktivDelelinje}
\title{FiktivDelelinje}
\format{
\preformatted{Simple feature collection with 1956 features and 1 field
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -94276.32 ymin: 6436015 xmax: 1119065 ymax: 7956336
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 1,956 x 2
   oppdateringsdato                                                     geometry
 * <date>                                                       <LINESTRING [m]>
 1 NA                                             (1e+06 7866606, 1e+06 7869714)
 2 NA                                             (1e+06 7866387, 1e+06 7866606)
 3 NA                                             (1e+06 7865264, 1e+06 7866387)
 4 NA                                             (1e+06 7865261, 1e+06 7865264)
 5 NA                                             (8e+05 7875160, 8e+05 7877324)
 6 NA               (1e+06 7874608, 1e+06 7874640, 1e+06 7874930, 1e+06 7876809)
 7 NA                                             (8e+05 7874601, 8e+05 7875160)
 8 NA                                             (8e+05 7874095, 8e+05 7874601)
 9 NA                                             (1e+06 7873838, 1e+06 7874608)
10 NA                                             (8e+05 7873663, 8e+05 7874095)
# ... with 1,946 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
FiktivDelelinje
}
\description{
FiktivDelelinje
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
