% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Fylkesgrense}
\alias{Fylkesgrense}
\title{Fylkesgrense}
\format{
\preformatted{Simple feature collection with 158 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -86015.82 ymin: 6466105 xmax: 627600.6 ymax: 7721400
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 158 x 4
   målemetode nøyaktighet oppdateringsdato                                     geometry
 *      <int>       <int> <date>                                       <LINESTRING [m]>
 1         64       50000 2020-12-14       (193105.8 6929798, 191705.8 6929024, 190961~
 2         64       50000 2020-12-14       (122837.3 6770942, 126146.2 6775329, 127136~
 3         64       50000 2020-12-14       (108149.8 6769548, 110025.3 6772727, 112927~
 4         64       50000 2020-12-14               (106502.3 6766939, 108149.8 6769548)
 5         64       50000 2020-12-14       (89184.93 6750237, 97935.87 6756949, 101612~
 6         64       50000 2020-12-14       (101558.2 6731476, 101428.2 6731564, 98540.~
 7         64       50000 2020-12-14       (98029.68 6719704, 98305.46 6721214, 98718.~
 8         64       50000 2020-12-14       (82868.22 6686184, 89079.01 6694104, 91406.~
 9         64       50000 2020-12-14       (82868.22 6686184, 82667.45 6685916, 81441.~
10         64       50000 2020-12-14       (38386.01 6609345, 40247.99 6612469, 40389.~
# ... with 148 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000AdministrativeOmrader_GML.gml}
}
\usage{
Fylkesgrense
}
\description{
Fylkesgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
