% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Grunnlinje}
\alias{Grunnlinje}
\title{Grunnlinje}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 1 feature and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -77243.1 ymin: 6448316 xmax: 1103784 ymax: 7940463
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 1 × 4
  oppdateringsdato målemetode nøyaktighet                                         grense
* <date>           <chr>            <int>                               <LINESTRING [m]>
1 2017-01-05       gen              50000 (1103784 7821829, 1101314 7848250, 1098543 78…
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000AdministrativeOmrader_GML.gml}
}
\usage{
Grunnlinje
}
\description{
Grunnlinje
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
