% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Gruve}
\alias{Gruve}
\title{Gruve}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 21 features and 3 fields
Geometry type: POINT
Dimension:     XY
Bounding box:  xmin: 191104 ymin: 6605222 xmax: 828437.1 ymax: 7778854
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 21 × 4
   målemetode nøyaktighet oppdateringsdato           posisjon
 *      <int>       <int> <date>                  <POINT [m]>
 1         64       50000 2017-01-05         (191681 6638002)
 2         64       50000 2017-01-05         (193989 6629555)
 3         64       50000 2017-01-05         (191104 6625863)
 4         64       50000 2017-01-05         (195586 6625863)
 5         64       50000 2017-01-05         (205578 6605222)
 6         64       50000 2017-01-05         (753103 7758172)
 7         64       50000 2017-01-05       (731320.4 7711147)
 8         64       50000 2017-01-05         (541894 7448584)
 9         64       50000 2017-01-05         (548195 7447393)
10         64       50000 2022-02-22       (828437.1 7778854)
# ℹ 11 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000BygningerOgAnlegg_GML.gml}
}
\usage{
Gruve
}
\description{
Gruve
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
