% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Hoydekurve}
\alias{Hoydekurve}
\title{Hoydekurve}
\format{
\preformatted{Simple feature collection with 19046 features and 5 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -55410 ymin: 6461757 xmax: 1114222 ymax: 7939626
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 19,046 x 6
   målemetode nøyaktighet medium oppdateringsdato høyde                        geometry
 *      <int>       <int> <chr>  <date>           <int>                <LINESTRING [m]>
 1         64       50000 T      1993-01-01         200 (1110182 7792918, 1109927 7792~
 2         64       50000 T      1993-01-01         100 (1080767 7800434, 1080965 7800~
 3         64       50000 T      1993-01-01         200 (1029347 7798160, 1029384 7798~
 4         64       50000 T      1993-01-01        1400 (104505.4 6876294, 104016 6876~
 5         64       50000 T      1993-01-01         200 (1055090 7775424, 1055340 7776~
 6         64       50000 T      1993-01-01         200 (1044296 7742663, 1044443 7742~
 7         64       50000 T      1993-01-01         100 (832045.3 7902584, 832046 7902~
 8         64       50000 T      1993-01-01         800 (559687.3 7436503, 559687.3 74~
 9         64       50000 T      1993-01-01        1400 (543314.7 7417128, 543314.7 74~
10         64       50000 T      1993-01-01        1200 (542968.5 7416437, 542968.5 74~
# ... with 19,036 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Hoyde_GML.gml}
}
\usage{
Hoydekurve
}
\description{
Høydekurve
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
