% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Hoydelag}
\alias{Hoydelag}
\title{Hoydelag}
\format{
\preformatted{Simple feature collection with 6326 features and 3 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -76208 ymin: 6450245 xmax: 1114945 ymax: 7939986
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 6,326 x 4
   oppdateringsdato makshøyde minhøyde                                         geometry
 * <date>               <int>    <int>                                    <POLYGON [m]>
 1 NA                    1200      600 ((904730 7801720, 904366 7801906, 904350 780207~
 2 NA                     600        0 ((680100.9 7800000, 677512.5 7800000, 677499 78~
 3 NA                     600        0 ((802837 7800952, 802640 7801335, 802617 780149~
 4 NA                     600        0 ((795078 7800000, 794745.7 7800000, 793880.1 78~
 5 NA                     600        0 ((657522 7800786, 657336 7801041, 657514 780121~
 6 NA                    1200      600 ((674054.5 7800000, 673202.6 7800000, 672940 78~
 7 NA                    1200      600 ((674923 7799948, 674421 7799798, 674292.9 7800~
 8 NA                    1200      600 ((835508.6 7800000, 835474 7799804, 834791 7799~
 9 NA                    1200      600 ((787929.2 7800000, 787436.3 7800000, 787611 78~
10 NA                     600        0 ((669757.1 7800000, 669702.6 7800000, 669736 78~
# ... with 6,316 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Hoyde_GML.gml}
}
\usage{
Hoydelag
}
\description{
Høydelag
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
