% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Industriomrade}
\alias{Industriomrade}
\title{Industriomrade}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 9 features and 3 fields
Geometry type: GEOMETRY
Dimension:     XY
Bounding box:  xmin: -55990.37 ymin: 6561408 xmax: 193508.3 ymax: 7046847
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 9 × 4
  oppdateringsdato målemetode nøyaktighet                                       geometry
* <date>                <int>       <int>                                 <GEOMETRY [m]>
1 2017-01-05               NA          NA MULTIPOLYGON (((193496.5 6565053, 193496 6565…
2 2017-01-05               NA          NA MULTIPOLYGON (((-40272.28 6782600, -40486 678…
3 2017-01-05               NA          NA MULTIPOLYGON (((-34485 6566338, -34720.13 656…
4 2017-01-05               NA          NA MULTIPOLYGON (((-38482 6609268, -38726.73 660…
5 2017-01-05               64       50000                         POINT (185163 7046847)
6 2017-01-05               64       50000                      POINT (-55990.37 6756067)
7 2017-01-05               64       50000                      POINT (-49932.91 6615034)
8 2017-01-05               64       50000                         POINT (190768 6562894)
9 2017-01-05               64       50000                         POINT (192058 6561408)
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
Industriomrade
}
\description{
Industriområde
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
