% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Innsjo}
\alias{Innsjo}
\title{Innsjo}
\format{
\preformatted{Simple feature collection with 3713 features and 2 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -54837 ymin: 6454936 xmax: 1111698 ymax: 7932555
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 3,713 x 3
   oppdateringsdato høyde                                                      geometry
 * <date>           <int>                                                 <POLYGON [m]>
 1 NA                 282 ((868487 7886343, 868370 7885910, 868471 7884716, 868399 788~
 2 NA                 179 ((887228 7885357, 886996 7885191, 886871 7884918, 886190 788~
 3 NA                 229 ((1021888 7884770, 1022112 7884466, 1022029 7884053, 1022246~
 4 NA                 149 ((883473 7885051, 883772 7884626, 883781 7884514, 883573 788~
 5 NA                 284 ((876886 7881714, 877156 7881989, 877286 7882650, 877468 788~
 6 NA                   6 ((854109 7881033, 853838 7880766, 853372 7880773, 852528 788~
 7 NA                  96 ((882241 7879134, 882311 7879297, 882649 7879527, 882808 787~
 8 NA                  42 ((965535 7930419, 965459 7930714, 965484 7931236, 965675 793~
 9 NA                   2 ((896794 7926270, 896889 7925965, 896422 7925634, 895603 792~
10 NA                 193 ((967381 7919988, 967128 7920220, 967047 7921191, 966906 792~
# ... with 3,703 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
Innsjo
}
\description{
Innsjø
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
