% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{InnsjoElvSperre}
\alias{InnsjoElvSperre}
\title{InnsjoElvSperre}
\format{
\preformatted{Simple feature collection with 22 features and 1 field
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 187700 ymin: 6575898 xmax: 1081162 ymax: 7771894
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 22 x 2
   oppdateringsdato                                            geometry
 * <date>                                              <LINESTRING [m]>
 1 2001-08-01                        (1081162 7771894, 1080993 7771753)
 2 2000-08-03                        (1072906 7765039, 1073109 7765076)
 3 2002-09-15       (1061965 7732954, 1062088 7733027, 1062091 7733027)
 4 1993-01-01                          (252376 6787082, 252560 6786912)
 5 2002-09-15                          (220285 6657320, 220412 6657235)
 6 2000-08-09                          (220855 6658206, 221092 6658018)
 7 2000-08-09                          (283181 6617402, 283322 6617295)
 8 2000-08-09                          (283978 6618337, 283882 6618448)
 9 2000-08-09                          (287438 6619581, 287409 6619803)
10 2002-09-15                          (273271 6577419, 273451 6577390)
# ... with 12 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
InnsjoElvSperre
}
\description{
InnsjøElvSperre
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
