% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Innsjokant}
\alias{Innsjokant}
\title{Innsjokant}
\format{
\preformatted{Simple feature collection with 7447 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -54837 ymin: 6454936 xmax: 1111698 ymax: 7932555
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 7,447 x 4
   målemetode nøyaktighet oppdateringsdato                                     geometry
 *      <int>       <int> <date>                                       <LINESTRING [m]>
 1         64       50000 1993-01-01       (830226 7867083, 829755 7866617, 829518 786~
 2         64       50000 1993-01-01       (966844 7866190, 966784 7866855, 966866 786~
 3         64       50000 1993-01-01       (964711 7865755, 964855 7865816, 965207 786~
 4         64       50000 1993-01-01       (994722 7865722, 994793 7867373, 995009 786~
 5         64       50000 1993-01-01       (956406 7865386, 955744 7865360, 955427 786~
 6         64       50000 1993-01-01       (875643 7865315, 875873 7865375, 875892 786~
 7         64       50000 1993-01-01       (865217 7872790, 865123 7872856, 866154 787~
 8         64       50000 1993-01-01       (871856 7872466, 872042 7872478, 872192 787~
 9         64       50000 1993-01-01       (988516 7872795, 988316 7872062, 988326 787~
10         64       50000 1992-12-31       (913958 7871644, 914140 7871790, 914491 787~
# ... with 7,437 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
Innsjokant
}
\description{
Innsjøkant
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
