% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{KantUtsnitt}
\alias{KantUtsnitt}
\title{KantUtsnitt}
\format{
\preformatted{Simple feature collection with 35 features and 1 field
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -99551.21 ymin: 6426052 xmax: 1121927 ymax: 7962739
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 35 x 2
   oppdateringsdato                                                            geometry
 * <date>                                                              <LINESTRING [m]>
 1 NA               (1e+06 7953601, 1007278 7950541, 1024145 7943449, 1042544 7938949,~
 2 NA               (1100000 7914538, 1109353 7909773, 1111054 7908706, 1112269 790780~
 3 NA               (9e+05 7956336, 914357.9 7957961, 927892.2 7959494, 951360.7 79621~
 4 NA               (8e+05 7921683, 809725.8 7929806, 811518.1 7931062, 813250.5 79320~
 5 NA               (774095.8 7900000, 782924.6 7907393, 796042.7 7918378, 8e+05 79216~
 6 NA               (7e+05 7852549, 701859.7 7853461, 726452.4 7865516, 744871 7874546~
 7 NA               (1119065 7900000, 1119670 7898840, 1120435 7896983, 1120921 789548~
 8 NA               (614333.1 7800000, 619273.9 7807759, 620492.7 7809343, 621154.4 78~
 9 NA               (6e+05 7777212, 606182.7 7787101, 609670.6 7792678, 614333.1 78000~
10 NA               (514317.9 7700000, 527306.2 7710050, 528110.2 7710643, 529536.9 77~
# ... with 25 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000KantUtsnitt_GML.gml}
}
\usage{
KantUtsnitt
}
\description{
KantUtsnitt
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
