% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Kommune}
\alias{Kommune}
\title{Kommune}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 359 features and 3 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -99551.21 ymin: 6426048 xmax: 1121941 ymax: 7962744
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 359 × 4
   oppdateringsdato kommunenummer navn                                            område
 * <date>           <chr>         <chr>                                    <POLYGON [m]>
 1 2023-12-18       1547          Aukra     ((83320.17 7010814, 72502.8 7030219, 63609.…
 2 2023-12-18       3110          Hvaler    ((265465.6 6562560, 261956.8 6562978, 25956…
 3 2023-12-18       1127          Randaberg ((-37004.41 6580926, -40878.87 6586533, -43…
 4 2023-12-18       1144          Kvitsøy   ((-40878.87 6586533, -42633.16 6593088, -50…
 5 2023-12-18       3212          Nesodden  ((258753.6 6645449, 257089.1 6646506, 25335…
 6 2023-12-18       1106          Haugesund ((-66131.86 6638129, -86015.82 6648235, -88…
 7 2023-12-18       4625          Austevoll ((-39434.61 6710972, -41600.44 6712732, -47…
 8 2023-12-18       1818          Herøy     ((355053.2 7359689, 334712.4 7369892, 32416…
 9 2023-12-18       4627          Askøy     ((-49560.77 6763448, -51197.2 6764298, -516…
10 2023-12-18       5020          Osen      ((263052.3 7172626, 247755.5 7188752, 22815…
# ℹ 349 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000AdministrativeOmrader_GML.gml}
}
\usage{
Kommune
}
\description{
Kommune
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
