% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Kommunegrense}
\alias{Kommunegrense}
\title{Kommunegrense}
\format{
\preformatted{Simple feature collection with 867 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -92079.54 ymin: 6428945 xmax: 1111593 ymax: 7962288
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 867 x 4
   målemetode nøyaktighet oppdateringsdato                                     geometry
 *      <int>       <int> <date>                                       <LINESTRING [m]>
 1         64       50000 2020-12-14       (505985.2 7582162, 505997.2 7582222, 506004~
 2         64       50000 2020-12-14       (299994.4 6770126, 301013.7 6770392, 302689~
 3         64       50000 2020-12-14       (299994.4 6770126, 301020.7 6768822, 302065~
 4         64       50000 2020-12-14       (299994.4 6770126, 299921.7 6770207, 298029~
 5         64       50000 2020-12-14       (40577.78 6547572, 41743.64 6547099, 42213.~
 6         64       50000 2020-12-14       (40577.78 6547572, 40700.49 6547728, 41051.~
 7         64       50000 2020-12-14       (204547.2 6759740, 203860.1 6759367, 201895~
 8         64       50000 2020-12-14       (204547.2 6759740, 204591.8 6759022, 204589~
 9         64       50000 2020-12-14       (294612.4 6629746, 294574.1 6629578, 294557~
10         64       50000 2020-12-14       (294612.4 6629746, 294522.7 6629696, 294449~
# ... with 857 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000AdministrativeOmrader_GML.gml}
}
\usage{
Kommunegrense
}
\description{
Kommunegrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
