% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Kystkontur}
\alias{Kystkontur}
\title{Kystkontur}
\format{
\preformatted{Simple feature collection with 6713 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -76208 ymin: 6450245 xmax: 1105197 ymax: 7939986
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 6,713 x 4
   målemetode nøyaktighet oppdateringsdato                                     geometry
 *      <int>       <int> <date>                                       <LINESTRING [m]>
 1         64       50000 2001-08-01       (902487 7865963, 903053.3 7866784, 903238.3~
 2         64       50000 1993-01-01       (976706 7865961, 976057 7865890, 975937 786~
 3         64       50000 1992-12-31       (938859 7865346, 938634 7865431, 938600 786~
 4         64       50000 1993-01-01       (985334 7869249, 985141 7869102, 984923 786~
 5         64       50000 1993-01-01       (993619 7875506, 993685 7875740, 993824 787~
 6         64       50000 1992-12-31       (898637 7876251, 898820 7876462, 899327 787~
 7         64       50000 2001-08-01       (8e+05 7875160, 799887 7875292, 799910 7875~
 8         64       50000 2001-08-01       (8e+05 7874601, 800081 7874648, 800212 7874~
 9         64       50000 1993-01-01       (999704 7874550, 999947 7875014, 999905 787~
10         64       50000 2001-08-01       (8e+05 7874095, 799976 7874121, 799808 7874~
# ... with 6,703 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
Kystkontur
}
\description{
Kystkontur
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
