% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Lufthavn}
\alias{Lufthavn}
\title{Lufthavn}
\format{
\preformatted{Simple feature collection with 55 features and 9 fields
Geometry type: POINT
Dimension:     XY
Bounding box:  xmin: -57805.33 ymin: 6472154 xmax: 1095819 ymax: 7929664
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 55 x 10
   målemetode nøyaktighet oppdateringsdato iataKode icaoKode lufthavneier  lufthavntype
 *      <int>       <int> <date>           <chr>    <chr>    <chr>         <chr>       
 1         64       50000 2011-11-29       NTB      ENNO     Notodden luf~ L           
 2         64       50000 2011-11-29       HFT      ENHF     Avinor AS     L           
 3         64       50000 2011-03-15       HAA      ENHK     Avinor AS     L           
 4         64       50000 2011-03-31       LKN      ENLK     Avinor AS     L           
 5         64       50000 2011-03-22       HOV      ENOV     Avinor AS     L           
 6         64       50000 2011-05-16       BJF      ENBS     Avinor AS     L           
 7         64       50000 2011-11-29       SVJ      ENSH     Avinor AS     L           
 8         64       50000 2011-03-25       ANX      ENAN     FLO Base And~ L           
 9         64       50000 2011-11-29       HVG      ENHV     Avinor AS     L           
10         64       50000 2011-03-31       SVG      ENZV     Avinor AS     L           
# ... with 45 more rows, and 3 more variables: navn <chr>, trafikktype <chr>,
#   geometry <POINT [m]>
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
Lufthavn
}
\description{
Lufthavn
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
