% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{LuftledningLH}
\alias{LuftledningLH}
\title{LuftledningLH}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 218 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -50520.24 ymin: 6463494 xmax: 1080873 ymax: 7853550
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 218 × 4
   oppdateringsdato målemetode nøyaktighet                                   senterlinje
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 2022-10-06       gen              50000 (40578 6913969, 40658.4 6914162, 40686.98 69…
 2 2022-10-06       gen              50000 (167545.1 6967736, 167476.1 6967746, 167200.…
 3 2022-10-06       gen              50000 (167545.1 6967736, 167645.7 6968030, 167618.…
 4 2022-10-06       gen              50000 (23793.71 6490067, 23857.86 6490294, 24067.1…
 5 2022-10-06       gen              50000 (20650.94 6531747, 20399.61 6531359, 20231.1…
 6 2022-10-06       gen              50000 (190049.3 6557412, 190847.9 6557867, 190984.…
 7 2022-10-03       gen              50000 (252754.6 6785839, 252897.9 6786041, 253287.…
 8 2022-10-06       gen              50000 (38600.53 6542374, 38672.77 6543065, 38738.5…
 9 2022-10-06       gen              50000 (54450.05 6600403, 52422.45 6600795, 52138.6…
10 2022-10-03       gen              50000 (189614.5 6583435, 189396.8 6583658, 189737.…
# ℹ 208 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000BygningerOgAnlegg_GML.gml}
}
\usage{
LuftledningLH
}
\description{
LuftledningLH
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
