% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Myr}
\alias{Myr}
\title{Myr}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 2864 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -58988.99 ymin: 6468671 xmax: 1114125 ymax: 7920789
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 2,864 × 2
   oppdateringsdato                                                               område
 * <date>                                                                  <POLYGON [m]>
 1 2017-01-05       ((1094181 7877371, 1094744 7878028, 1095456 7877976, 1095539 787822…
 2 2017-01-05       ((906559 7880343, 906762 7880351, 906938 7880462, 907461.1 7881142,…
 3 2017-01-05       ((895923 7714082, 895407 7714307, 894851 7714136, 894692.7 7714006,…
 4 2017-01-05       ((916960 7878352, 917651 7879363, 917539 7879523, 917372 7879590, 9…
 5 2017-01-05       ((924175 7900000, 924183 7900287, 923936 7900317, 923272 7900144, 9…
 6 2017-01-05       ((973162 7916613, 973261.2 7916634, 973263 7916634, 973826 7916843,…
 7 2017-01-05       ((919098.8 7906979, 919493.8 7907154, 919492.8 7907506, 919297.8 79…
 8 2017-01-05       ((828119.5 7900000, 828229 7900159, 828262 7900348, 827895 7900651,…
 9 2017-01-05       ((834861 7778528, 835010 7778641, 835414 7778946, 835009 7779303, 8…
10 2017-01-05       ((1088649 7875529, 1088903 7875532, 1089434 7875963, 1089570 787618…
# ℹ 2,854 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
Myr
}
\description{
Myr
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
