% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N5000-package.R
\docType{package}
\name{N1000-package}
\alias{N1000}
\alias{N1000-package}
\title{N1000: Kartdata}
\source{
\itemize{
\item \url{https://kartkatalog.geonorge.no/metadata/kartverket/n1000-kartdata/aee42bb6-d0e9-4d70-86fe-6ea76c381055}
}
}
\description{
Kartdata fra \href{https://kartverket.no/}{Kartverket}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/hmalmedal/N1000}
  \item Report bugs at \url{https://github.com/hmalmedal/N1000/issues}
}

}
\author{
\strong{Maintainer}: Håkon Malmedal \email{hmalmedal@gmail.com}

Other contributors:
\itemize{
  \item Kartverket [copyright holder]
}

}
\keyword{package}
