% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Naturverngrense}
\alias{Naturverngrense}
\title{Naturverngrense}
\format{
\preformatted{Simple feature collection with 53 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 1183.26 ymin: 6472365 xmax: 1088054 ymax: 7893205
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 53 x 4
   målemetode nøyaktighet oppdateringsdato                                     geometry
 *      <int>       <int> <date>                                       <LINESTRING [m]>
 1         64       50000 2016-02-01       (494981.1 7472050, 493139 7471776, 493107.7~
 2         64       50000 2016-02-01       (509820 7483182, 510618.5 7484554, 510464.8~
 3         64       50000 2016-02-01       (642141.4 7608935, 638731.5 7611581, 636996~
 4         64       50000 2016-02-01       (92235.33 6740841, 92009.31 6742561, 90324.~
 5         64       50000 2016-02-01       (28219.27 6691623, 25983.92 6690898, 25870.~
 6         64       50000 2016-02-01       (585497 7677287, 585690 7678220, 585370 767~
 7         64       50000 2016-02-01       (533613.6 7475526, 534740.7 7479364, 535951~
 8         64       50000 2016-02-01       (753705.9 7706934, 756208.5 7709680, 758654~
 9         64       50000 2016-02-01       (271280.2 6550636, 271264.8 6550607, 271216~
10         64       50000 2016-02-01       (863873 7658542, 863904.8 7658597, 864120.4~
# ... with 43 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Restriksjonsomrader_GML.gml}
}
\usage{
Naturverngrense
}
\description{
Naturverngrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
