% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Naturvernomrade}
\alias{Naturvernomrade}
\title{Naturvernomrade}
\format{
\preformatted{Simple feature collection with 42 features and 4 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: 1183.26 ymin: 6472365 xmax: 1088054 ymax: 7893205
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 42 x 5
   oppdateringsdato navn    vernedato  verneform                               geometry
 * <date>           <chr>   <date>     <chr>                              <POLYGON [m]>
 1 2014-01-28       Sjunkh~ 2010-02-05 NP        ((494981.1 7472050, 493139 7471776, 4~
 2 2014-01-28       Rohkun~ 2011-02-25 NP        ((642141.4 7608935, 638731.5 7611581,~
 3 2014-01-28       Hallin~ 2006-12-22 NP        ((92235.33 6740841, 92009.31 6742561,~
 4 2014-01-28       Folgef~ 2005-04-29 NP        ((28219.27 6691623, 25983.92 6690898,~
 5 2014-01-28       Ånderd~ 2004-06-04 NP        ((585497 7677287, 585690 7678220, 585~
 6 2014-01-28       Rago    1971-01-22 NP        ((533613.6 7475526, 534740.7 7479364,~
 7 2014-01-28       Reisa   1986-11-28 NP        ((753705.9 7706934, 756208.5 7709680,~
 8 2014-01-28       Ytre H~ 2009-06-26 NP        ((271280.2 6550636, 271264.8 6550607,~
 9 2014-01-28       Øvre A~ 1975-12-19 NP        ((863873 7658542, 863904.8 7658597, 8~
10 2014-01-28       Øvre P~ 2003-08-29 NP        ((1043856 7739675, 1045491 7739623, 1~
# ... with 32 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Restriksjonsomrader_GML.gml}
}
\usage{
Naturvernomrade
}
\description{
Naturvernområde
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
