% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Navigasjonsinstallasjon}
\alias{Navigasjonsinstallasjon}
\title{Navigasjonsinstallasjon}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 106 features and 4 fields
Geometry type: POINT
Dimension:     XY
Bounding box:  xmin: -75270 ymin: 6449852 xmax: 1098990 ymax: 7939496
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 106 × 5
   målemetode nøyaktighet oppdateringsdato fyrlistenummer           posisjon
 *      <int>       <int> <date>                    <int>        <POINT [m]>
 1         64       50000 2017-01-05                20900   (253016 6639164)
 2         64       50000 2017-01-05                31600 (248095.2 6585792)
 3         64       50000 2017-01-05                33500 (250038.4 6565999)
 4         64       50000 2017-01-05                53100   (261511 6564372)
 5         64       50000 2017-01-05                  100   (243201 6551650)
 6         64       50000 2017-01-05                44200   (198903 6550952)
 7         64       50000 2017-01-05                 5500   (271787 6548674)
 8         64       50000 2017-01-05                  200   (257732 6547330)
 9         64       50000 2017-01-05                41700   (221145 6546755)
10         64       50000 2017-01-05                43500   (208731 6544222)
# ℹ 96 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000BygningerOgAnlegg_GML.gml}
}
\usage{
Navigasjonsinstallasjon
}
\description{
Navigasjonsinstallasjon
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
