% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Passasjerferjestrekning}
\alias{Passasjerferjestrekning}
\title{Passasjerferjestrekning}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 25 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 59772 ymin: 6544205 xmax: 445470 ymax: 7214682
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 25 × 4
   målemetode nøyaktighet oppdateringsdato                                   senterlinje
 *      <int>       <int> <date>                                        <LINESTRING [m]>
 1         64       50000 2017-01-05       (121388.1 6691851, 121388.9 6691848, 121778.…
 2         64       50000 2017-01-05       (104307.6 6711695, 104313.5 6711700, 104440.…
 3         64       50000 2017-01-05       (343821.1 6917814, 343819.5 6917813, 343641.…
 4         64       50000 2017-01-05       (337579.3 6903838, 337589.7 6903833, 338321.…
 5         64       50000 2017-01-05       (125379 6649884, 125411.2 6649978, 125414.9 …
 6         64       50000 2017-01-05       (339120.1 6924195, 339107.7 6924078, 339136.…
 7         64       50000 2017-01-05       (170571.7 6833873, 168819.4 6833780, 167963.…
 8         64       50000 2017-01-05       (161164.3 6833755, 160886.8 6833311, 159291.…
 9         64       50000 2017-01-05       (340267.1 6912784, 340279 6912775, 340560.9 …
10         64       50000 2017-01-05       (97840.54 6720300, 98333.72 6719846, 98467.7…
# ℹ 15 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Samferdsel_GML.gml}
}
\usage{
Passasjerferjestrekning
}
\description{
Passasjerferjestrekning
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
