% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Riksgrense}
\alias{Riksgrense}
\title{Riksgrense}
\format{
\preformatted{Simple feature collection with 36 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 295815.3 ymin: 6531564 xmax: 1115056 ymax: 7828443
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 36 x 4
   målemetode nøyaktighet oppdateringsdato                                     geometry
 *      <int>       <int> <date>                                       <LINESTRING [m]>
 1         64       50000 2020-12-14       (1008977 7804953, 1011539 7802961, 1012505 ~
 2         64       50000 2020-12-14                 (1008765 7808532, 1008977 7804953)
 3         64       50000 2020-12-14       (931415.7 7785646, 931769.7 7785963, 932947~
 4         64       50000 2020-12-14       (747182.8 7701826, 750698 7701151, 753776.2~
 5         64       50000 2020-12-14       (738347 7692234, 740578.4 7695439, 744627.3~
 6         64       50000 2020-12-14       (701756.5 7667791, 705043.6 7668334, 711360~
 7         64       50000 2020-12-14       (701248.5 7613149, 703307.9 7615067, 705613~
 8         64       50000 2020-12-14       (627600.6 7605923, 631602.5 7607940, 634530~
 9         64       50000 2020-12-14       (352618.2 6652316, 353733.1 6654174, 354829~
10         64       50000 2020-12-14       (916230.4 7677406, 916183.9 7677460, 916173~
# ... with 26 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000AdministrativeOmrader_GML.gml}
}
\usage{
Riksgrense
}
\description{
Riksgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
