% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Skog}
\alias{Skog}
\title{Skog}
\format{
\preformatted{Simple feature collection with 2884 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -56133 ymin: 6450729 xmax: 1114945 ymax: 7921538
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 2,884 x 2
   oppdateringsdato                                                            geometry
 * <date>                                                                 <POLYGON [m]>
 1 NA               ((951956 7911719, 951448 7911740, 951210 7911917, 951265 7912322, ~
 2 NA               ((953052 7910658, 952745 7910721, 953309 7911391, 953486.8 7911424~
 3 NA               ((956097 7907628, 956029 7907648, 955884 7907691, 955805.8 7907765~
 4 NA               ((951364.2 7900000, 951721.3 7900000, 951742.7 7900000, 951732 789~
 5 NA               ((1e+06 7873838, 999825 7873929, 999617 7874038, 999521 7874200, 9~
 6 NA               ((977579 7907649, 977449 7907683, 977167 7907557, 976937 7907333, ~
 7 NA               ((951742.7 7900000, 951721.3 7900000, 951364.2 7900000, 950102 790~
 8 NA               ((908181 7869131, 907246 7869589, 907025 7869906, 906862 7870384, ~
 9 NA               ((852935.1 7890473, 852941.2 7890467, 853085.9 7890572, 853222.6 7~
10 NA               ((913655 7876619, 912987 7876051, 912868 7875555, 912155 7875457, ~
# ... with 2,874 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
Skog
}
\description{
Skog
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
