% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Skytefelt}
\alias{Skytefelt}
\title{Skytefelt}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 15 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: 45802.74 ymin: 6623336 xmax: 1075266 ymax: 7849300
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 15 × 2
   navn                                                                           område
 * <chr>                                                                   <POLYGON [m]>
 1 Giskås SØF                ((353391.7 7105440, 353385.3 7105443, 353378.6 7105446, 35…
 2 Halkavarre SØF            ((891351 7781073, 891371.8 7781081, 891392.4 7781090, 8919…
 3 Haltdalen SØF             ((300770.8 6983702, 300765.9 6983699, 300761.7 6983694, 30…
 4 Leksdal SØF               ((307400.6 7039261, 307365.5 7039289, 306607 7039882, 3065…
 5 Mauken Blåtind SØF        ((657122.1 7685400, 657115.6 7685402, 657109.3 7685405, 65…
 6 Rødsmoen SØF              ((311901.8 6793141, 311348.6 6793650, 311299.3 6793695, 31…
 7 Mjølfjell og Brandset SØF ((55578.52 6771648, 55572.83 6771649, 55566.66 6771650, 55…
 8 Hengsvann SØF             ((190108.1 6624248, 190111.2 6624255, 190113.5 6624261, 19…
 9 Nyborgmoen skytefelt      ((1012967 7848130, 1012963 7848136, 1012958 7848141, 10129…
10 Regionfelt Østlandet      ((311508.2 6791933, 311670 6791548, 311693.1 6791492, 3117…
11 Setermoen SØF             ((636042.3 7641219, 636037.4 7641225, 636032.2 7641230, 63…
12 Høybuktmoen SØF           ((1072221 7809587, 1072214 7809586, 1072207 7809585, 10722…
13 Terningmoen SØF           ((311141.4 6755406, 311134.6 6755406, 311127.8 6755404, 31…
14 Vaterholmen               ((348766.8 7065037, 348767.9 7065044, 348768.8 7065052, 34…
15 Steinsjøen SØF            ((284500.2 6715578, 284498.8 6715576, 284497.8 6715575, 28…
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Restriksjonsomrader_GML.gml}
}
\usage{
Skytefelt
}
\description{
Skytefelt
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
