% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Skytefeltgrense}
\alias{Skytefeltgrense}
\title{Skytefeltgrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 16 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 45802.74 ymin: 6623336 xmax: 1075266 ymax: 7849300
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 16 × 4
   oppdateringsdato målemetode nøyaktighet                                        grense
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 NA               gen              50000 (284500.2 6715578, 284500.5 6715581, 284500.…
 2 NA               gen              50000 (348766.8 7065037, 348765.2 7065030, 348763.…
 3 NA               gen              50000 (311141.4 6755406, 311148.4 6755406, 311155.…
 4 NA               gen              50000 (636042.3 7641219, 636047.2 7641214, 636051.…
 5 NA               gen              50000 (311901.8 6793141, 312650.2 6793777, 312673 …
 6 NA               gen              50000 (1012967 7848130, 1012970 7848124, 1012972 7…
 7 NA               gen              50000 (55578.52 6771648, 55583.64 6771645, 55588.2…
 8 NA               gen              50000 (311508.2 6791933, 311335.1 6791667, 311293.…
 9 NA               gen              50000          (311901.8 6793141, 311508.2 6791933)
10 NA               gen              50000 (657122.1 7685400, 657128.8 7685399, 657135.…
11 NA               gen              50000 (307400.6 7039261, 307417.7 7039248, 307485.…
12 NA               gen              50000 (1072221 7809587, 1072229 7809588, 1072236 7…
13 NA               gen              50000 (190108.1 6624248, 190104.3 6624243, 190099.…
14 NA               gen              50000 (300770.8 6983702, 300776.3 6983705, 300782.…
15 NA               gen              50000 (891351 7781073, 891330 7781065, 891301.8 77…
16 NA               gen              50000 (353391.7 7105440, 353397.7 7105436, 353403.…
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Restriksjonsomrader_GML.gml}
}
\usage{
Skytefeltgrense
}
\description{
Skytefeltgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
