% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{SnoIsbre}
\alias{SnoIsbre}
\title{SnoIsbre}
\format{
\preformatted{Simple feature collection with 118 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -5664 ymin: 6649920 xmax: 811950.2 ymax: 7837518
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 118 x 2
   oppdateringsdato                                                            geometry
 * <date>                                                                 <POLYGON [m]>
 1 NA               ((763484.8 7808343, 763592 7807866, 763663.6 7807711, 763718 78075~
 2 NA               ((768702.6 7802638, 768983.5 7802422, 769119 7802318, 769834.2 780~
 3 NA               ((691542.8 7706433, 691539 7706370, 691607 7706202, 691787.5 77061~
 4 NA               ((477654.5 7401440, 477646.2 7401324, 477117 7401432, 476978 74014~
 5 NA               ((452352 7413498, 451766.1 7413959, 451483.7 7414180, 451408 74142~
 6 NA               ((669816 7636044, 670315.7 7635120, 670181.6 7634905, 670118.2 763~
 7 NA               ((474806.6 7397547, 474655.5 7397421, 474168.3 7397479, 473796.4 7~
 8 NA               ((471712.7 7432991, 471587.2 7432723, 471661.8 7431856, 471690 743~
 9 NA               ((591095.2 7572683, 591255 7572388, 591319.1 7572410, 591892.1 757~
10 NA               ((113807.8 6938262, 113739 6938068, 113449.5 6937597, 112988 69368~
# ... with 108 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
SnoIsbre
}
\description{
SnøIsbre
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
