% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{StedsnavnTekst}
\alias{StedsnavnTekst}
\title{StedsnavnTekst}
\format{
\preformatted{Simple feature collection with 5545 features and 18 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -76547.7 ymin: 6447274 xmax: 1110222 ymax: 7941194
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 5,545 x 19
   streng oppdateringsdato datafangstdato tekstReferanseP~ tekstReferanseP~ tegnavstand
 * <chr>  <date>           <date>                    <int>            <int>       <int>
 1 Skoro~ 2017-11-10       2017-01-06                    0                0           0
 2 Holand 2017-11-10       2017-01-06                    0                0           0
 3 Sørli  2017-11-10       2017-01-06                    0                0           0
 4 Stugg~ 2017-11-10       2017-01-06                    0                1           0
 5 Láhpo~ 2017-11-10       2017-01-06                    0                0           0
 6 Bergs~ 2017-11-10       2017-01-06                    0                0           0
 7 Vanna~ 2017-11-10       2017-01-06                    0                0           0
 8 Kinna~ 2017-11-10       2017-01-06                    0                2           0
 9 Mehamn 2017-11-10       2017-01-06                    0                2           0
10 lufth~ 2017-11-10       2017-01-06                    0                2           0
# ... with 5,535 more rows, and 13 more variables: skriftkode <int>, skrifttype <chr>,
#   referansemålestokk <int>, fulltekst <chr>, navneobjekthovedgruppe <int>,
#   navneobjektgruppe <int>, navneobjekttype <int>, språk <chr>,
#   stedsnavnsnummer <int>, språkprioritering <chr>, skrivemåtenummer <int>,
#   stedsnummer <int>, geometry <LINESTRING [m]>
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Stedsnavn_GML.gml}
}
\usage{
StedsnavnTekst
}
\description{
StedsnavnTekst
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
