% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Sti}
\alias{Sti}
\title{Sti}
\format{
\preformatted{Simple feature collection with 904 features and 4 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -38821 ymin: 6527264 xmax: 1022648 ymax: 7899446
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 904 x 5
   målemetode nøyaktighet oppdateringsdato merking                             geometry
 *      <int>       <int> <date>           <chr>                       <LINESTRING [m]>
 1         64       50000 2005-06-01       JA      (760049 7715757, 761258.9 7715493, ~
 2         64       50000 2002-09-15       JA      (755209 7718900, 755143 7718899, 75~
 3         64       50000 2002-09-15       JA      (775455 7704871, 775480.9 7704839, ~
 4         64       50000 2002-09-15       JA      (810550 7682223, 810782.5 7682333, ~
 5         64       50000 2002-09-15       JA      (811940 7681316, 811899.4 7681203, ~
 6         64       50000 2002-09-15       JA      (703898 7665254, 703428 7663204, 70~
 7         64       50000 2002-09-15       JA      (705539 7642898, 706005.6 7642486, ~
 8         64       50000 2013-04-04       JA      (184989.5 7008291, 184963.5 7008383~
 9         64       50000 2002-09-15       JA      (515892 7353961, 515794 7354181, 51~
10         64       50000 2002-09-15       JA      (147806.9 6921673, 147742 6922620, ~
# ... with 894 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Samferdsel_GML.gml}
}
\usage{
Sti
}
\description{
Sti
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
