% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Terrengpunkt}
\alias{Terrengpunkt}
\title{Terrengpunkt}
\format{
\preformatted{Simple feature collection with 3472 features and 5 fields
Geometry type: POINT
Dimension:     XY
Bounding box:  xmin: -73523 ymin: 6455520 xmax: 1108784 ymax: 7937526
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 3,472 x 6
   målemetode nøyaktighet medium oppdateringsdato høyde         geometry
 *      <int>       <int> <chr>  <date>           <int>      <POINT [m]>
 1         64       50000 T      2012-11-08         766 (509047 7448933)
 2         64       50000 T      2012-11-08         572 (472733 7448909)
 3         64       50000 T      2012-11-08        1908 (559686 7448415)
 4         64       50000 T      1999-12-10        1822 (555202 7447982)
 5         64       50000 T      2012-11-08         991 (525032 7447785)
 6         64       50000 T      2012-11-08         644 (514969 7445684)
 7         64       50000 T      2012-11-08        1006 (530853 7445336)
 8         64       50000 T      1980-07-01        1073 (482968 7444636)
 9         64       50000 T      2012-11-08         225 (452015 7444113)
10         64       50000 T      2012-11-08        1063 (536440 7442682)
# ... with 3,462 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Hoyde_GML.gml}
}
\usage{
Terrengpunkt
}
\description{
Terrengpunkt
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
