% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Territorialgrense}
\alias{Territorialgrense}
\title{Territorialgrense}
\format{
\preformatted{Simple feature collection with 77 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -99551.21 ymin: 6426048 xmax: 1121941 ymax: 7962744
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 77 x 4
   målemetode nøyaktighet oppdateringsdato                                     geometry
 *      <int>       <int> <date>                                       <LINESTRING [m]>
 1         64       50000 2020-12-14       (1093794 7917699, 1108469 7910248, 1109353 ~
 2         64       50000 2020-12-14       (1043616 7938687, 1056895 7935439, 1057864 ~
 3         64       50000 2020-12-14       (1007278 7950541, 1024145 7943449, 1043616 ~
 4         64       50000 2020-12-14       (953401.7 7962288, 972585.4 7962738, 973255~
 5         64       50000 2020-12-14       (927892.2 7959494, 951360.7 7962151, 952358~
 6         64       50000 2020-12-14       (852906 7943866, 874031.9 7952542, 874966.4~
 7         64       50000 2020-12-14       (796042.5 7918378, 808944.4 7929181, 809725~
 8         64       50000 2020-12-14       (764795.7 7891994, 770308.5 7896821, 770685~
 9         64       50000 2020-12-14       (726452.5 7865517, 744871 7874546, 764795.7~
10         64       50000 2020-12-14               (701859.7 7853461, 726452.5 7865517)
# ... with 67 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000AdministrativeOmrader_GML.gml}
}
\usage{
Territorialgrense
}
\description{
Territorialgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
