% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Tettbebyggelse}
\alias{Tettbebyggelse}
\title{Tettbebyggelse}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 1191 features and 2 fields (with 989 geometries empty)
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -58727 ymin: 6455392 xmax: 1098753 ymax: 7913676
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 1,191 × 3
   målemetode nøyaktighet                                                         område
 * <chr>            <int>                                                  <POLYGON [m]>
 1 <NA>                NA ((-37958.21 6572598, -37605 6572309, -37478.01 6572415, -3744…
 2 <NA>                NA ((241118 6574939, 240679 6575380, 240660.4 6575423, 240308 65…
 3 <NA>                NA ((-22670 6515816, -22570 6515963, -22569 6516450, -22624.32 6…
 4 <NA>                NA ((198606.6 6551140, 198603 6551270, 198573 6552314, 198441 65…
 5 <NA>                NA ((253791.1 6598132, 253449.5 6598105, 253305 6598093, 253301 …
 6 <NA>                NA ((265122.1 6568601, 265339 6568873, 265580 6569282, 265786 65…
 7 <NA>                NA ((-29867 6575198, -29689.99 6575501, -29438 6575703, -29514 6…
 8 <NA>                NA ((241670 6574728, 241817 6574777, 241912 6575034, 241908 6575…
 9 <NA>                NA ((141695.6 6498750, 141996 6498340, 142263.8 6498266, 142317 …
10 <NA>                NA ((-50717.51 6623334, -50819 6623494, -51005 6623676, -50959 6…
# ℹ 1,181 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
Tettbebyggelse
}
\description{
Tettbebyggelse
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
