% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Veglenke}
\alias{Veglenke}
\title{Veglenke}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 14466 features and 10 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -74500 ymin: 6452103 xmax: 1112920 ymax: 7939337
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 14,466 × 11
   medium oppdateringsdato målemetode nøyaktighet typeVeg  vegkategori vegfase vegnummer
 * <chr>  <date>           <chr>            <int> <chr>    <chr>       <chr>       <int>
 1 T      2017-01-05       gen              50000 enkelBi… K           V              NA
 2 T      2017-01-05       gen              50000 enkelBi… P           V              NA
 3 T      2017-01-05       gen              50000 enkelBi… K           V              NA
 4 T      2017-01-05       gen              50000 enkelBi… P           V              NA
 5 T      2017-01-05       gen              50000 enkelBi… K           V              NA
 6 T      2017-01-05       gen              50000 enkelBi… P           V              NA
 7 T      2017-01-05       gen              50000 enkelBi… P           V              NA
 8 T      2017-01-05       gen              50000 enkelBi… K           V              NA
 9 T      2017-01-05       gen              50000 enkelBi… K           V              NA
10 T      2017-01-05       gen              50000 enkelBi… K           V              NA
# ℹ 14,456 more rows
# ℹ 3 more variables: motorvegtype <chr>, rutemerking <chr>,
#   senterlinje <LINESTRING [m]>
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Samferdsel_GML.gml}
}
\usage{
Veglenke
}
\description{
Veglenke
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
