% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{VegSenterlinje}
\alias{VegSenterlinje}
\title{VegSenterlinje}
\format{
\preformatted{Simple feature collection with 13037 features and 8 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -74500 ymin: 6451272 xmax: 1112920 ymax: 7939337
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 13,037 x 9
   målemetode nøyaktighet medium oppdateringsdato vegkategori vegstatus vegnummer
 *      <int>       <int> <chr>  <date>           <chr>       <chr>         <int>
 1         64       50000 T      2006-07-01       K           V                NA
 2         64       50000 T      2006-07-01       P           V                NA
 3         64       50000 T      2002-09-15       F           V                NA
 4         64       50000 T      2001-08-01       K           V                NA
 5         64       50000 T      2002-09-15       K           V                NA
 6         64       50000 T      2015-04-15       P           V                NA
 7         64       50000 T      2001-08-01       K           V                NA
 8         64       50000 T      2004-07-01       P           V                NA
 9         64       50000 T      2001-08-01       F           V                NA
10         64       50000 T      2014-04-22       P           V                NA
# ... with 13,027 more rows, and 2 more variables: motorvegtype <chr>,
#   geometry <LINESTRING [m]>
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Samferdsel_GML.gml}
}
\usage{
VegSenterlinje
}
\description{
VegSenterlinje
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
