% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{AnnenBatrute}
\alias{AnnenBatrute}
\title{AnnenBatrute}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 53 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -59426.35 ymin: 6734699 xmax: 1102790 ymax: 7941234
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 53 × 4
   oppdateringsdato målemetode nøyaktighet                                   senterlinje
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 2017-01-05       gen              50000            (1012611 7922331, 1011518 7921259)
 2 2017-01-05       gen              50000 (966769.2 7938989, 970232.1 7940345, 971955.…
 3 2017-01-05       gen              50000 (1012611 7922331, 1013836 7923732, 1014391 7…
 4 2017-01-05       gen              50000 (1043402 7908353, 1044139 7910827, 1044706 7…
 5 2017-01-05       gen              50000 (1097176 7888490, 1098501 7888274, 1099453 7…
 6 2017-01-05       gen              50000 (1088500 7850137, 1084103 7847118, 1078877 7…
 7 2017-01-05       gen              50000 (1075218 7823496, 1075284 7827537, 1075937 7…
 8 2017-01-05       gen              50000 (1055103 7839692, 1057570 7838621, 1063510 7…
 9 2017-01-05       gen              50000 (962099 7934023, 962907.3 7935718, 963479.5 …
10 2017-01-05       gen              50000 (962099 7934023, 961746 7936037, 961077.8 79…
# ℹ 43 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Samferdsel_GML.gml}
}
\usage{
AnnenBatrute
}
\description{
AnnenBåtrute
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
