% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{ApentOmrade}
\alias{ApentOmrade}
\title{ApentOmrade}
\format{
\preformatted{Simple feature collection with 5850 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -76208 ymin: 6450245 xmax: 1114893 ymax: 7939986
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 5,850 x 2
   oppdateringsdato                                                            geometry
 * <date>                                                                 <POLYGON [m]>
 1 NA               ((948111 7885686, 946490.7 7886156, 946257.3 7886423, 946193 78866~
 2 NA               ((818028 7886086, 817650 7886227, 817603 7886373, 817427 7886364, ~
 3 NA               ((804854 7881733, 805438 7881521, 805768 7881689, 806126 7881392, ~
 4 NA               ((1033313 7841945, 1032762 7841886, 1032496 7841925, 1032281 78420~
 5 NA               ((895999 7713275, 896107 7713812, 896047.4 7713899, 895948.3 77140~
 6 NA               ((794309 7878043, 793620 7878062, 793343 7878235, 792522 7878161, ~
 7 NA               ((8e+05 7877324, 8e+05 7875160, 799887 7875292, 799910 7875902, 79~
 8 NA               ((1069626 7900000, 1069220 7900000, 1069506 7900080, 1069626 79000~
 9 NA               ((821940.7 7900000, 823342.8 7900000, 823534.6 7900000, 824604.8 7~
10 NA               ((871370 7924455, 871194 7924481, 870562 7924874, 870240 7924882, ~
# ... with 5,840 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
ApentOmrade
}
\description{
ÅpentOmråde
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
