% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{ApentOmrade}
\alias{ApentOmrade}
\title{ApentOmrade}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 5850 features and 0 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -76208 ymin: 6450245 xmax: 1114893 ymax: 7939986
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 5,850 × 1
                                                                                  område
 *                                                                         <POLYGON [m]>
 1 ((948111 7885686, 948163.1 7885767, 948228.7 7885869, 948238 7885884, 948622 7886045…
 2 ((818028 7886086, 818062 7886206, 817951 7886553, 817709 7886767, 817517 7886725, 81…
 3 ((804854 7881733, 804692 7882145, 803777 7882843, 803795 7883268, 803374 7883837, 80…
 4 ((1033313 7841945, 1032390 7842118, 1031396 7842808, 1031381 7842819, 1030967 784297…
 5 ((895999 7713275, 896257 7713345, 896842 7713082, 896944 7713781, 897373 7714396, 89…
 6 ((794309 7878043, 795077 7878512, 795148 7878679, 794966 7879297, 794526 7879777, 79…
 7 ((8e+05 7877324, 799640 7877623, 799345 7878249, 799076 7878523, 798817 7878304, 798…
 8                ((1069626 7900000, 1069506 7900080, 1069220 7900000, 1069626 7900000))
 9 ((821940.7 7900000, 821739 7899667, 821898 7899071, 822681 7899310, 823014 7899529, …
10 ((871370 7924455, 871679 7924575, 871862 7924761, 871704 7925306, 871667 7925504, 87…
# ℹ 5,840 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
ApentOmrade
}
\description{
ÅpentOmråde
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
