% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Arealbrukgrense}
\alias{Arealbrukgrense}
\title{Arealbrukgrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 14953 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -58988.99 ymin: 6454505 xmax: 1114893 ymax: 7921538
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 14,953 × 4
   oppdateringsdato målemetode nøyaktighet                                        grense
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 2017-01-05       gen              50000 (905013 7868638, 905270 7868613, 905932 7868…
 2 2017-01-05       gen              50000 (905013 7868638, 904948 7868478, 904980 7867…
 3 2017-01-05       gen              50000 (1e+06 7869714, 1000424 7869137, 1000550 786…
 4 2017-01-05       gen              50000 (986717 7867608, 986512 7867436, 986440.1 78…
 5 2017-01-05       gen              50000 (991429 7866476, 991394 7866734, 991249.5 78…
 6 2017-01-05       gen              50000 (1e+06 7866606, 999926 7866536, 1e+06 786638…
 7 2017-01-05       gen              50000 (829228 7866857, 829164 7867184, 829032 7867…
 8 2017-01-05       gen              50000 (1036278 7865928, 1036892 7866288, 1037208 7…
 9 2017-01-05       gen              50000 (1e+06 7866387, 1000139 7866108, 1e+06 78652…
10 2017-01-05       gen              50000 (872246 7865227, 872476 7865294, 872683 7865…
# ℹ 14,943 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
Arealbrukgrense
}
\description{
Arealbrukgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
