% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Bane}
\alias{Bane}
\title{Bane}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 304 features and 4 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -42446 ymin: 6466305 xmax: 1076833 ymax: 7806691
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 304 × 5
   medium oppdateringsdato målemetode nøyaktighet                            senterlinje
 * <chr>  <date>           <chr>            <int>                       <LINESTRING [m]>
 1 T      2022-09-26       gen              50000 (198691.8 6707898, 199082.5 6707999, …
 2 U      2022-09-26       gen              50000 (201287 6707536, 201345.1 6707515, 20…
 3 U      2022-09-26       gen              50000 (207100 6695272, 207298.8 6695265, 20…
 4 T      2022-09-26       gen              50000 (196541 6565927, 196710 6565829, 1967…
 5 T      2022-09-26       gen              50000 (334977 6676088, 335390 6676253, 3354…
 6 U      2022-09-26       gen              50000 (-1031.17 6753578, -962.78 6753599, -…
 7 T      2022-09-26       gen              50000         (76900 6750992, 77205 6750843)
 8 T      2022-09-26       gen              50000 (101817 6734511, 102069.7 6734567, 10…
 9 T      2022-09-26       gen              50000 (202320.3 6707154, 202536.1 6707043, …
10 T      2022-09-26       gen              50000 (209419 6695191, 209649.1 6695064, 21…
# ℹ 294 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Samferdsel_GML.gml}
}
\usage{
Bane
}
\description{
Bane
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
