% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Barmarksloype}
\alias{Barmarksloype}
\title{Barmarksloype}
\format{
\preformatted{Simple feature collection with 60 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 797893.2 ymin: 7650834 xmax: 1066724 ymax: 7873456
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 60 x 4
   målemetode nøyaktighet oppdateringsdato                                     geometry
 *      <int>       <int> <date>                                       <LINESTRING [m]>
 1         64       50000 2015-09-16       (1058259 7846052, 1058418 7846930, 1058449 ~
 2         64       50000 2011-01-24       (838375 7769214, 838796.8 7769149, 839237.7~
 3         64       50000 2011-01-24       (908692 7867390, 908631.8 7867583, 908577.7~
 4         64       50000 2011-01-24       (943577.1 7873259, 943734 7873331, 944078 7~
 5         64       50000 2011-01-24       (843746.7 7764239, 843762.7 7764222, 844098~
 6         64       50000 2011-01-24       (851211.9 7756127, 850794.7 7756375, 850444~
 7         64       50000 2012-04-12       (1011624 7831045, 1011544 7831102, 1011457 ~
 8         64       50000 2011-01-24               (833431.6 7779698, 833577.6 7779595)
 9         64       50000 2011-01-24       (825584.6 7686748, 825069.2 7687199, 824849~
10         64       50000 2011-01-24       (869174.8 7823140, 869388.5 7823135, 869787~
# ... with 50 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Samferdsel_GML.gml}
}
\usage{
Barmarksloype
}
\description{
Barmarksløype
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
