% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Bilferjestrekning}
\alias{Bilferjestrekning}
\title{Bilferjestrekning}
\format{
\preformatted{Simple feature collection with 282 features and 6 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -74407.93 ymin: 6484917 xmax: 850374.3 ymax: 7904854
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 282 x 7
   målemetode nøyaktighet medium oppdateringsdato vegkategori vegnummer
 *      <int>       <int> <chr>  <date>           <chr>           <int>
 1         64       50000 T      2001-08-01       F                  NA
 2         64       50000 T      2016-03-01       F                  17
 3         64       50000 T      2001-08-01       F                  NA
 4         64       50000 T      2010-07-06       F                 839
 5         64       50000 T      2016-03-01       F                 839
 6         64       50000 T      2010-07-06       F                  17
 7         64       50000 T      2007-10-15       F                  NA
 8         64       50000 T      2007-10-15       F                  NA
 9         64       50000 T      2007-10-15       F                  NA
10         64       50000 T      2010-07-06       F                 469
# ... with 272 more rows, and 1 more variable: geometry <LINESTRING [m]>
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Samferdsel_GML.gml}
}
\usage{
Bilferjestrekning
}
\description{
Bilferjestrekning
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
