% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Bygning}
\alias{Bygning}
\title{Bygning}
\format{
\preformatted{Simple feature collection with 3237 features and 7 fields
Geometry type: POINT
Dimension:     XY
Bounding box:  xmin: -74711.91 ymin: 6452445 xmax: 1107651 ymax: 7930848
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 3,237 x 8
   målemetode nøyaktighet oppdateringsdato bygningskategori betjeningsgrad hytteeier
 *      <int>       <int> <date>                      <int> <chr>              <int>
 1         64       50000 2017-12-11                     56 U                      1
 2         64       50000 2017-12-11                     56 U                      1
 3         64       50000 2017-12-11                     56 U                      1
 4         64       50000 2017-12-11                     56 U                      1
 5         64       50000 2017-12-11                     56 U                      1
 6         64       50000 2017-12-11                     56 U                      1
 7         64       50000 2017-12-11                     56 U                      1
 8         64       50000 2017-12-11                     56 U                      2
 9         64       50000 2017-12-11                     56 U                      2
10         64       50000 2017-12-11                     56 U                      2
# ... with 3,227 more rows, and 2 more variables: navn <chr>, geometry <POINT [m]>
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000BygningerOgAnlegg_GML.gml}
}
\usage{
Bygning
}
\description{
Bygning
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
