% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Bygning}
\alias{Bygning}
\title{Bygning}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 2701 features and 8 fields
Geometry type: POINT
Dimension:     XY
Bounding box:  xmin: -74711.91 ymin: 6454575 xmax: 1107651 ymax: 7930848
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 2,701 × 9
   oppdateringsdato målemetode nøyaktighet bygningskategori betjeningsgrad hytteeier
 * <date>           <chr>            <int>            <int> <chr>              <int>
 1 2017-11-30       gen              50000               56 Selvbetjent            1
 2 2017-11-30       gen              50000               56 Selvbetjent            1
 3 2017-11-30       gen              50000               56 Betjent                1
 4 2017-11-30       gen              50000               56 Betjent                1
 5 2017-11-30       gen              50000               56 Betjent                1
 6 2017-11-30       gen              50000               56 Betjent                1
 7 2017-11-30       gen              50000               56 Selvbetjent            1
 8 2017-11-30       gen              50000               56 Selvbetjent            1
 9 2017-11-30       gen              50000               56 Selvbetjent            1
10 2017-11-30       gen              50000               56 Selvbetjent            1
# ℹ 2,691 more rows
# ℹ 3 more variables: tilgjengelighet <chr>, navn <chr>, posisjon <POINT [m]>
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000BygningerOgAnlegg_GML.gml}
}
\usage{
Bygning
}
\description{
Bygning
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
