% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Dataavgrensning}
\alias{Dataavgrensning}
\title{Dataavgrensning}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 588 features and 1 field
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 247627.9 ymin: 6531544 xmax: 1114945 ymax: 7828329
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 588 × 2
   oppdateringsdato                                                               grense
 * <date>                                                               <LINESTRING [m]>
 1 2017-01-05                                       (974864.2 7819430, 975096.5 7821747)
 2 2017-01-05                                       (971128.4 7816274, 973341.3 7817037)
 3 2017-01-05       (367165.8 6715570, 366730.6 6716276, 366710.8 6716309, 365703.7 671…
 4 2017-01-05                                       (970066.1 7814397, 971109.1 7816252)
 5 2017-01-05       (1084659 7800880, 1084666 7800878, 1085922 7800429, 1086703 7800000)
 6 2017-01-05                     (971109.1 7816252, 971119.6 7816271, 971128.4 7816274)
 7 2017-01-05                                         (1080262 7772197, 1081122 7771294)
 8 2017-01-05                                         (1047696 7749804, 1047597 7749594)
 9 2017-01-05       (1059377 7752892, 1059305 7752753, 1059495 7751427, 1059431 7750983…
10 2017-01-05                                       (949176.8 7804990, 949867.5 7805176)
# ℹ 578 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
Dataavgrensning
}
\description{
Dataavgrensning
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
