% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{ElvBekk}
\alias{ElvBekk}
\title{ElvBekk}
\format{
\preformatted{Simple feature collection with 9054 features and 4 fields
Geometry type: GEOMETRY
Dimension:     XY
Bounding box:  xmin: -58946 ymin: 6455880 xmax: 1114945 ymax: 7935034
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 9,054 x 5
   oppdateringsdato vannBredde målemetode nøyaktighet                          geometry
 * <date>                <int>      <int>       <int>                     <POLYGON [m]>
 1 NA                        6         NA          NA ((818880 7782886, 818829.5 77827~
 2 NA                        6         NA          NA ((293243.2 6700000, 292950.7 670~
 3 NA                        6         NA          NA ((255124 6797955, 255270 6797310~
 4 NA                        6         NA          NA ((248192 6805572, 248163 6805528~
 5 NA                        6         NA          NA ((221092 6658018, 220871 6657835~
 6 NA                        6         NA          NA ((293088 6699272, 293267 6698452~
 7 NA                        6         NA          NA ((3e+05 6665875, 3e+05 6666076, ~
 8 NA                        6         NA          NA ((283322 6617295, 283309.2 66172~
 9 2007-10-04                6         NA          NA ((1062063 7733070, 1062091 77330~
10 NA                        6         NA          NA ((273451 6577390, 273447 6577290~
# ... with 9,044 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
ElvBekk
}
\description{
ElvBekk
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
