% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{ElvBekk}
\alias{ElvBekk}
\title{ElvBekk}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 9027 features and 1 field
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -58946 ymin: 6455880 xmax: 1114945 ymax: 7935034
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 9,027 × 2
   vannbredde                                                                senterlinje
 *      <int>                                                           <LINESTRING [m]>
 1          3 (515482 7531651, 515381 7531844, 515053 7531588, 514695 7531516, 514482 7…
 2          3                         (415054 7531135, 415194.3 7531113, 415199 7531112)
 3          3                                           (502054 7531015, 503639 7531658)
 4          3 (558253 7530908, 558487.2 7530511, 558506 7530479, 558376.5 7530230, 5582…
 5          3                           (499029 7529054, 498613 7528194, 498584 7528134)
 6          3 (564989 7529179, 564254 7529319, 564003.8 7529137, 563893.8 7529058, 5637…
 7          4                           (534631 7526323, 535466 7527747, 535641 7528585)
 8          4 (521218 7526076, 521221.5 7526104, 521270 7526489, 520946 7526723, 518969…
 9          3 (508499 7525602, 508185 7525831, 508821.3 7526480, 509107 7526772, 509152…
10          3 (553802 7525518, 554029.9 7526039, 554063.3 7526115, 554163 7526343, 5549…
# ℹ 9,017 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
ElvBekk
}
\description{
ElvBekk
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
