% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{ElvBekkKant}
\alias{ElvBekkKant}
\title{ElvBekkKant}
\format{
\preformatted{Simple feature collection with 386 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 187462 ymin: 6566414 xmax: 1081802 ymax: 7873955
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 386 x 4
   målemetode nøyaktighet oppdateringsdato                                     geometry
 *      <int>       <int> <date>                                       <LINESTRING [m]>
 1         64       50000 1993-01-01       (991429 7866476, 991951 7867260, 992167 786~
 2         64       50000 1998-01-01       (990838 7865195, 990862 7865333, 990970 786~
 3         64       50000 1993-01-01                   (993185 7873440, 993516 7873955)
 4         64       50000 1993-01-01       (994613 7873673, 994441 7873143, 994343 787~
 5         64       50000 1998-01-01       (992147 7869620, 992394 7870151, 992497 787~
 6         64       50000 1993-01-01       (994592 7871079, 994683 7870068, 994623 786~
 7         64       50000 1993-01-01       (990673 7860838, 990656 7861206, 991060 786~
 8         64       50000 1993-01-01       (993037 7850008, 993461 7849161, 993609 784~
 9         64       50000 1993-01-01       (991721 7858486, 991540 7858660, 991493 785~
10         64       50000 1993-01-01       (995919 7840607, 996022 7840223, 996242 783~
# ... with 376 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
ElvBekkKant
}
\description{
ElvBekkKant
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
