% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{FiktivDelelinje}
\alias{FiktivDelelinje}
\title{FiktivDelelinje}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 1956 features and 1 field
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -94276.32 ymin: 6436015 xmax: 1119065 ymax: 7956336
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 1,956 × 2
   oppdateringsdato                                                       grense
 * <date>                                                       <LINESTRING [m]>
 1 2017-01-05                                     (1e+06 7866606, 1e+06 7869714)
 2 2017-01-05                                     (1e+06 7866387, 1e+06 7866606)
 3 2017-01-05                                     (1e+06 7865264, 1e+06 7866387)
 4 2017-01-05                                     (1e+06 7865261, 1e+06 7865264)
 5 2017-01-05                                     (8e+05 7875160, 8e+05 7877324)
 6 2017-01-05       (1e+06 7874608, 1e+06 7874640, 1e+06 7874930, 1e+06 7876809)
 7 2017-01-05                                     (8e+05 7874601, 8e+05 7875160)
 8 2017-01-05                                     (8e+05 7874095, 8e+05 7874601)
 9 2017-01-05                               (882794.1 7900000, 885043.1 7900000)
10 2017-01-05                                     (1e+06 7873838, 1e+06 7874608)
# ℹ 1,946 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
FiktivDelelinje
}
\description{
FiktivDelelinje
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
