% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Fylkesgrense}
\alias{Fylkesgrense}
\title{Fylkesgrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 178 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -86015.82 ymin: 6466105 xmax: 805691.2 ymax: 7853461
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 178 × 4
   målemetode nøyaktighet oppdateringsdato                                        grense
 *      <int>       <int> <date>                                        <LINESTRING [m]>
 1         64       50000 2023-12-18       (199915.1 6544617, 200281.5 6540388, 208835 …
 2         64       50000 2023-12-18       (188287.4 7050695, 172271.5 7045714, 167403.…
 3         64       50000 2023-12-18       (135156.4 6913742, 131853.4 6913593, 128871.…
 4         64       50000 2023-12-18       (247009.1 6931442, 245356.9 6937394, 244610.…
 5         64       50000 2023-12-18       (623019.9 7613727, 623406.8 7613041, 624087.…
 6         64       50000 2023-12-18       (134985.3 6829516, 136125.5 6830011, 137352.…
 7         64       50000 2023-12-18                (435603.1 7220891, 437120.3 7221190)
 8         64       50000 2023-12-18       (-28106.39 6661595, -27211.11 6661362, -2592…
 9         64       50000 2023-12-18       (139140.6 6538162, 138892.6 6538532, 138781.…
10         64       50000 2023-12-18       (302615.9 6696022, 302783.7 6695747, 302852.…
# ℹ 168 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000AdministrativeOmrader_GML.gml}
}
\usage{
Fylkesgrense
}
\description{
Fylkesgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
