% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Fylkesgrense}
\alias{Fylkesgrense}
\title{Fylkesgrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 178 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -86015.82 ymin: 6466105 xmax: 805691.2 ymax: 7853461
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 178 × 4
   oppdateringsdato målemetode nøyaktighet                                        grense
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 2023-12-18       gen              50000 (302615.9 6696022, 302783.7 6695747, 302852.…
 2 2023-12-18       gen              50000 (122336.4 6915968, 121352.4 6915041, 117672.…
 3 2023-12-18       gen              50000 (139140.6 6538162, 138892.6 6538532, 138781.…
 4 2023-12-18       gen              50000 (-28106.39 6661595, -27211.11 6661362, -2592…
 5 2023-12-18       gen              50000 (174750.7 6652275, 174858.7 6652492, 175092.…
 6 2023-12-18       gen              50000 (101558.2 6731476, 101417.1 6730969, 101168.…
 7 2023-12-18       gen              50000 (213013.5 6615913, 213409.3 6615940, 214538.…
 8 2023-12-18       gen              50000 (353306.9 7219739, 353217.9 7219477, 354568.…
 9 2023-12-18       gen              50000 (777865.5 7678331, 778466.3 7678413, 782818.…
10 2023-12-18       gen              50000 (544620 7647160, 541220 7644590, 540270 7641…
# ℹ 168 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000AdministrativeOmrader_GML.gml}
}
\usage{
Fylkesgrense
}
\description{
Fylkesgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
