% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Gruve}
\alias{Gruve}
\title{Gruve}
\format{
\preformatted{Simple feature collection with 45 features and 3 fields
Geometry type: POINT
Dimension:     XY
Bounding box:  xmin: 3446 ymin: 6545199 xmax: 1077058 ymax: 7798105
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 45 x 4
   målemetode nøyaktighet oppdateringsdato         geometry
 *      <int>       <int> <date>                <POINT [m]>
 1         64       50000 2000-10-02       (192992 6643260)
 2         64       50000 1995-09-19       (197240 6639624)
 3         64       50000 2000-10-02       (191681 6638002)
 4         64       50000 2000-10-02       (204739 6637554)
 5         64       50000 1995-09-19       (208987 6637498)
 6         64       50000 1995-09-19       (193989 6629555)
 7         64       50000 2000-10-02       (191104 6625863)
 8         64       50000 2002-09-15       (195586 6625863)
 9         64       50000 1995-09-19       (205578 6605222)
10         64       50000 2000-09-28       (206198 6555786)
# ... with 35 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000BygningerOgAnlegg_GML.gml}
}
\usage{
Gruve
}
\description{
Gruve
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
