% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{HavElvSperre}
\alias{HavElvSperre}
\title{HavElvSperre}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 11 features and 1 field
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 192821 ymin: 6566414 xmax: 1080701 ymax: 7873955
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 11 × 2
   oppdateringsdato                                                               grense
 * <date>                                                               <LINESTRING [m]>
 1 2017-01-05                                           (994613 7873673, 993516 7873955)
 2 2017-01-05                                         (1080614 7802291, 1080701 7802465)
 3 2017-01-05                                           (818880 7782886, 818702 7783127)
 4 2017-01-05                                           (332664 7153221, 332707 7152781)
 5 2017-01-05                           (638039 7689267, 639389 7689587, 639580 7689587)
 6 2017-01-05                         (417657 7302472, 417654.7 7302382, 417649 7302163)
 7 2017-01-05                                           (325699 7080015, 325715 7079814)
 8 2017-01-05                                           (229672 6632846, 229639 6632987)
 9 2017-01-05       (192821 6567513, 192897.5 6567387, 192897.5 6567387, 192897.7 65673…
10 2017-01-05           (268594 6566414, 268681 6566472, 269091 6566745, 269200 6566818)
11 2017-01-05           (265025 6568479, 265090 6568443, 265266 6568345, 265355 6568295)
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
HavElvSperre
}
\description{
HavElvSperre
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
