% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Havflate}
\alias{Havflate}
\title{Havflate}
\format{
\preformatted{Simple feature collection with 209 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -99551.21 ymin: 6426052 xmax: 1121927 ymax: 7962739
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 209 x 2
   oppdateringsdato                                                            geometry
 * <date>                                                                 <POLYGON [m]>
 1 NA               ((820805.1 7900000, 821699 7900159, 822000 7900098, 821940.7 79000~
 2 NA               ((774095.8 7900000, 782924.6 7907393, 796042.7 7918378, 8e+05 7921~
 3 NA               ((1119065 7900000, 1100000 7900000, 1100000 7914538, 1109353 79097~
 4 NA               ((924921.7 7900000, 925828.6 7900000, 925762 7899870, 925381 78996~
 5 NA               ((975292.1 7900000, 975750.1 7900000, 975679 7899914, 975326.9 789~
 6 NA               ((1068820 7900000, 1069220 7900000, 1069024 7899945, 1068846 78999~
 7 NA               ((1069626 7900000, 1070599 7900000, 1070479 7899936, 1070228 78998~
 8 NA               ((1071690 7900000, 1072785 7900000, 1072573 7899948, 1071823 78999~
 9 NA               ((856420.3 7900000, 860539.8 7900000, 860502 7899871, 860107 78996~
10 NA               ((891060 7896829, 890934 7896785, 890709.8 7896800, 890585 7896808~
# ... with 199 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
Havflate
}
\description{
Havflate
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
