% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Havflate}
\alias{Havflate}
\title{Havflate}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 209 features and 0 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -99551.21 ymin: 6426052 xmax: 1121927 ymax: 7962739
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 209 × 1
                                                                                  område
 *                                                                         <POLYGON [m]>
 1 ((820805.1 7900000, 821940.7 7900000, 822000 7900098, 821699 7900159, 820805.1 79000…
 2 ((774095.8 7900000, 8e+05 7900000, 8e+05 7921683, 796042.7 7918378, 782924.6 7907393…
 3 ((1119065 7900000, 1118741 7900620, 1117657 7902310, 1116424 7903895, 1115306 790512…
 4 ((924921.7 7900000, 924764 7899744, 924900 7899393, 925164 7899324, 925381 7899630, …
 5 ((975292.1 7900000, 975299.7 7899892, 975309 7899760, 975326.9 7899767, 975679 78999…
 6 ((1068820 7900000, 1068846 7899965, 1069024 7899945, 1069220 7900000, 1068820 790000…
 7 ((1069626 7900000, 1069827 7899866, 1070228 7899859, 1070479 7899936, 1070599 790000…
 8 ((1071690 7900000, 1071823 7899932, 1072573 7899948, 1072785 7900000, 1071690 790000…
 9 ((856420.3 7900000, 856498 7899819, 856970 7899302, 857315 7898644, 857182 7898115, …
10 ((891060 7896829, 891269 7897101, 891375.8 7897651, 891182.5 7898021, 891318.9 78984…
# ℹ 199 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
Havflate
}
\description{
Havflate
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
