% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Hoydekurve}
\alias{Hoydekurve}
\title{Hoydekurve}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 19042 features and 5 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -55410 ymin: 6461757 xmax: 1114222 ymax: 7939626
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 19,042 × 6
   målemetode nøyaktighet medium oppdateringsdato høyde                      senterlinje
 *      <int>       <int> <chr>  <date>           <int>                 <LINESTRING [m]>
 1         64       50000 T      2017-01-05         200 (1110182 7792918, 1109927 77929…
 2         64       50000 T      2017-01-05         800 (559687.3 7436503, 559687.3 743…
 3         64       50000 T      2017-01-05        1400 (543314.7 7417128, 543314.7 741…
 4         64       50000 T      2017-01-05         100 (417538.2 7359144, 417507 73589…
 5         64       50000 I      2017-01-05        1200 (464311.3 7392200, 464316.6 739…
 6         64       50000 T      2017-01-05         200 (1055090 7775424, 1055340 77760…
 7         64       50000 T      2017-01-05         200 (1044296 7742663, 1044443 77425…
 8         64       50000 T      2017-01-05         100 (832045.3 7902584, 832046 79025…
 9         64       50000 T      2017-01-05        1200 (542968.5 7416437, 542968.5 741…
10         64       50000 T      2017-01-05        1200 (542036.6 7414579, 542024.4 741…
# ℹ 19,032 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Hoyde_GML.gml}
}
\usage{
Hoydekurve
}
\description{
Høydekurve
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
