% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Hoydekurve}
\alias{Hoydekurve}
\title{Hoydekurve}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 19042 features and 5 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -55410 ymin: 6461757 xmax: 1114222 ymax: 7939626
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 19,042 × 6
   medium oppdateringsdato målemetode nøyaktighet høyde                      senterlinje
 * <chr>  <date>           <chr>            <int> <int>                 <LINESTRING [m]>
 1 T      2017-01-05       gen              50000   200 (1110182 7792918, 1109927 77929…
 2 T      2017-01-05       gen              50000   800 (559687.3 7436503, 559687.3 743…
 3 T      2017-01-05       gen              50000  1400 (543314.7 7417128, 543314.7 741…
 4 T      2017-01-05       gen              50000   100 (417538.2 7359144, 417507 73589…
 5 I      2017-01-05       gen              50000  1400 (91545.15 6905364, 91587.94 690…
 6 T      2017-01-05       gen              50000   200 (1055090 7775424, 1055340 77760…
 7 T      2017-01-05       gen              50000   200 (1044296 7742663, 1044443 77425…
 8 T      2017-01-05       gen              50000   100 (832045.3 7902584, 832046 79025…
 9 T      2017-01-05       gen              50000  1200 (542968.5 7416437, 542968.5 741…
10 T      2017-01-05       gen              50000  1200 (542036.6 7414579, 542024.4 741…
# ℹ 19,032 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Hoyde_GML.gml}
}
\usage{
Hoydekurve
}
\description{
Høydekurve
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
