% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Hoydelag}
\alias{Hoydelag}
\title{Hoydelag}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 6327 features and 3 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -76208 ymin: 6450245 xmax: 1114945 ymax: 7939986
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 6,327 × 4
   oppdateringsdato makshøyde minhøyde                                            område
 * <date>               <int>    <int>                                     <POLYGON [m]>
 1 2017-01-05            1200      600 ((904730 7801720, 904875 7802083, 904685 7802230…
 2 2017-01-05             600        0 ((680100.9 7800000, 680117 7800073, 679965 78005…
 3 2017-01-05             600        0 ((802837 7800952, 803233 7800975, 803174 7801161…
 4 2017-01-05             600        0 ((795078 7800000, 794946 7800295, 794304 7800715…
 5 2017-01-05             600        0 ((657522 7800786, 657699 7800858, 657956 7801116…
 6 2017-01-05            1200      600 ((674054.5 7800000, 673462 7801132, 672940 78006…
 7 2017-01-05            1200      600 ((674923 7799948, 674958.1 7800000, 675101 78002…
 8 2017-01-05            1200      600 ((835508.6 7800000, 835590 7800461, 835511 78005…
 9 2017-01-05            1200      600 ((787929.2 7800000, 787852 7800121, 787611 78001…
10 2017-01-05             600        0 ((669757.1 7800000, 669736 7800039, 669702.6 780…
# ℹ 6,317 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Hoyde_GML.gml}
}
\usage{
Hoydelag
}
\description{
Høydelag
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
