% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Hoydelag}
\alias{Hoydelag}
\title{Hoydelag}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 6327 features and 2 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -76208 ymin: 6450245 xmax: 1114945 ymax: 7939986
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 6,327 × 3
   makshøyde minhøyde                                                             område
 *     <int>    <int>                                                      <POLYGON [m]>
 1      1200      600 ((904730 7801720, 904875 7802083, 904685 7802230, 904350 7802078,…
 2       600        0 ((680100.9 7800000, 680117 7800073, 679965 7800523, 679874 780161…
 3       600        0 ((802837 7800952, 803233 7800975, 803174 7801161, 802977 7801486,…
 4       600        0 ((795078 7800000, 794946 7800295, 794304 7800715, 794399 7801157,…
 5       600        0 ((657522 7800786, 657699 7800858, 657956 7801116, 657927 7801219,…
 6      1200      600 ((674054.5 7800000, 673462 7801132, 672940 7800653, 673202.6 7800…
 7      1200      600 ((674923 7799948, 674958.1 7800000, 675101 7800212, 674734 780037…
 8      1200      600 ((835508.6 7800000, 835590 7800461, 835511 7800596, 835384 780049…
 9      1200      600 ((787929.2 7800000, 787852 7800121, 787611 7800151, 787436.3 7800…
10       600        0 ((669757.1 7800000, 669736 7800039, 669702.6 7800000, 669757.1 78…
# ℹ 6,317 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Hoyde_GML.gml}
}
\usage{
Hoydelag
}
\description{
Høydelag
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
