% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Industriomrade}
\alias{Industriomrade}
\title{Industriomrade}
\format{
\preformatted{Simple feature collection with 9 features and 3 fields
Geometry type: GEOMETRY
Dimension:     XY
Bounding box:  xmin: -55990.37 ymin: 6561408 xmax: 193508.3 ymax: 7046847
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 9 x 4
  oppdateringsdato målemetode nøyaktighet                                      geometry
* <date>                <int>       <int>                                <GEOMETRY [m]>
1 2011-03-22               NA          NA MULTIPOLYGON (((-34485 6566338, -34453.38 65~
2 2011-03-22               NA          NA MULTIPOLYGON (((-38482 6609268, -38639 66092~
3 2011-03-22               NA          NA MULTIPOLYGON (((193496.5 6565053, 193508.3 6~
4 2011-03-22               NA          NA MULTIPOLYGON (((-40272.28 6782600, -40897.52~
5 2000-09-28               64       50000                        POINT (185163 7046847)
6 2004-07-01               64       50000                     POINT (-55990.37 6756067)
7 2004-07-01               64       50000                     POINT (-49932.91 6615034)
8 2004-07-01               64       50000                        POINT (190768 6562894)
9 2004-07-01               64       50000                        POINT (192058 6561408)
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
Industriomrade
}
\description{
Industriområde
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
