% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Industriomrade}
\alias{Industriomrade}
\title{Industriomrade}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 9 features and 2 fields
Geometry type: GEOMETRY
Dimension:     XY
Bounding box:  xmin: -55990.37 ymin: 6561408 xmax: 193508.3 ymax: 7046847
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 9 × 3
  målemetode nøyaktighet                                                        geometry
* <chr>            <int>                                                  <GEOMETRY [m]>
1 <NA>                NA POLYGON ((193496.5 6565053, 193496 6565052, 193410 6565126, 19…
2 <NA>                NA POLYGON ((-40272.28 6782600, -40486 6782759, -40368 6783426, -…
3 <NA>                NA POLYGON ((-34485 6566338, -34720.13 6566847, -34935.13 6566796…
4 <NA>                NA POLYGON ((-38482 6609268, -38726.73 6609874, -38734.39 6609873…
5 gen              50000                                          POINT (185163 7046847)
6 gen              50000                                       POINT (-55990.37 6756067)
7 gen              50000                                       POINT (-49932.91 6615034)
8 gen              50000                                          POINT (190768 6562894)
9 gen              50000                                          POINT (192058 6561408)
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
Industriomrade
}
\description{
Industriområde
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
