% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Innsjo}
\alias{Innsjo}
\title{Innsjo}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 3713 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -54837 ymin: 6454936 xmax: 1111698 ymax: 7932555
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 3,713 × 2
   høyde                                                                          område
 * <int>                                                                   <POLYGON [m]>
 1   282 ((868487 7886343, 868517 7887251, 868282 7887203, 868271 7886927, 867882 78865…
 2   179 ((887228 7885357, 887562 7886180, 887294 7885958, 887043 7885548, 886396 78851…
 3   229 ((1021888 7884770, 1021363 7885298, 1021085 7885349, 1020992 7885197, 1020964 …
 4   149 ((883473 7885051, 882667 7884835, 882478 7884396, 882075 7884278, 882043 78840…
 5   284 ((876886 7881714, 875825 7879857, 875771 7879510, 875930 7879384, 876209 78795…
 6     6 ((854109 7881033, 853344 7881456, 852552 7881293, 852468 7881194, 852528 78810…
 7    96 ((882241 7879134, 882049 7878943, 882483 7878282, 882994 7878156, 883114 78780…
 8    42 ((965535 7930419, 965640 7930538, 965617 7930723, 965977 7931021, 966161 79313…
 9     2 ((896794 7926270, 896585 7926639, 896356 7926618, 895547 7925854, 895290 79254…
10   193 ((967381 7919988, 967543 7920785, 967290 7920898, 967235 7921299, 967603 79224…
# ℹ 3,703 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
Innsjo
}
\description{
Innsjø
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
