% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Innsjokant}
\alias{Innsjokant}
\title{Innsjokant}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 7447 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -54837 ymin: 6454936 xmax: 1111698 ymax: 7932555
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 7,447 × 4
   oppdateringsdato målemetode nøyaktighet                                        grense
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 2017-01-05       gen              50000 (830226 7867083, 829755 7866617, 829518 7866…
 2 2017-01-05       gen              50000 (966844 7866190, 966784 7866855, 966866 7867…
 3 2017-01-05       gen              50000 (964711 7865755, 964855 7865816, 965207 7866…
 4 2017-01-05       gen              50000 (994722 7865722, 994793 7867373, 995009 7867…
 5 2017-01-05       gen              50000 (956406 7865386, 955744 7865360, 955427 7865…
 6 2017-01-05       gen              50000 (875643 7865315, 875873 7865375, 875892 7865…
 7 2017-01-05       gen              50000 (865217 7872790, 865123 7872856, 866154 7874…
 8 2017-01-05       gen              50000 (871856 7872466, 872042 7872478, 872192 7872…
 9 2017-01-05       gen              50000 (988516 7872795, 988316 7872062, 988326 7871…
10 2017-01-05       gen              50000 (913958 7871644, 914140 7871790, 914491 7872…
# ℹ 7,437 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
Innsjokant
}
\description{
Innsjøkant
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
