% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{KantUtsnitt}
\alias{KantUtsnitt}
\title{KantUtsnitt}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 35 features and 1 field
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -99551.21 ymin: 6426052 xmax: 1121927 ymax: 7962739
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 35 × 2
   oppdateringsdato                                                               grense
 * <date>                                                               <LINESTRING [m]>
 1 2017-01-05       (1e+06 7953601, 1007278 7950541, 1024145 7943449, 1042544 7938949, …
 2 2017-01-05       (1100000 7914538, 1109353 7909773, 1111054 7908706, 1112269 7907804…
 3 2017-01-05       (9e+05 7956336, 914357.9 7957961, 927892.2 7959494, 951360.7 796215…
 4 2017-01-05       (8e+05 7921683, 809725.8 7929806, 811518.1 7931062, 813250.5 793206…
 5 2017-01-05       (774095.8 7900000, 782924.6 7907393, 796042.7 7918378, 8e+05 792168…
 6 2017-01-05       (7e+05 7852549, 701859.7 7853461, 726452.4 7865516, 744871 7874546,…
 7 2017-01-05       (1119065 7900000, 1119670 7898840, 1120435 7896983, 1120921 7895480…
 8 2017-01-05       (614333.1 7800000, 619273.9 7807759, 620492.7 7809343, 621154.4 781…
 9 2017-01-05       (6e+05 7777212, 606182.7 7787101, 609670.6 7792678, 614333.1 780000…
10 2017-01-05       (514317.9 7700000, 527306.2 7710050, 528110.2 7710643, 529536.9 771…
# ℹ 25 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
KantUtsnitt
}
\description{
KantUtsnitt
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
