% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Kommune}
\alias{Kommune}
\title{Kommune}
\format{
\preformatted{Simple feature collection with 357 features and 3 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -99551.21 ymin: 6426048 xmax: 1121941 ymax: 7962744
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 357 x 4
   oppdateringsdato kommunenummer navn                                         geometry
 * <date>           <chr>         <chr>                                   <POLYGON [m]>
 1 2020-12-14       4227          Kvinesd~ ((41743.64 6547099, 42213.8 6547092, 42918.~
 2 2020-12-14       3811          Færder   ((237775.4 6578542, 238334.8 6577615, 23855~
 3 2020-12-14       1874          Moskenes ((414337.5 7551147, 416631.2 7549294, 41686~
 4 2020-12-14       1114          Bjerkre~ ((-8949.85 6550402, -7949.72 6550366, -7789~
 5 2020-12-14       4224          Åseral   ((65590.71 6551161, 65630.88 6551147, 65832~
 6 2020-12-14       3012          Aremark  ((309239.5 6582510, 311852.6 6582191, 31213~
 7 2020-12-14       3004          Fredrik~ ((270218.2 6583096, 270232.4 6583080, 27058~
 8 2020-12-14       1127          Randabe~ ((-37004.41 6580926, -33915.07 6579005, -34~
 9 2020-12-14       4206          Farsund  ((16887.22 6483470, 17423.53 6482581, 17719~
10 2020-12-14       3822          Nissedal ((124366.8 6588275, 125274.8 6588074, 12620~
# ... with 347 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000AdministrativeOmrader_GML.gml}
}
\usage{
Kommune
}
\description{
Kommune
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
