% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Kommune}
\alias{Kommune}
\title{Kommune}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 359 features and 4 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -99551.21 ymin: 6426048 xmax: 1121941 ymax: 7962744
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 359 × 5
   kommunenummer kommunenavn fylkesnummer fylkesnavn                              område
 * <chr>         <chr>       <chr>        <chr>                            <POLYGON [m]>
 1 1560          Tingvoll    15           Møre og Romsdal ((152677.3 7018684, 148348.8 …
 2 1547          Aukra       15           Møre og Romsdal ((83320.17 7010814, 72502.8 7…
 3 5618          Måsøy       56           Finnmark        ((861361.7 7923284, 852906 79…
 4 1103          Stavanger   11           Rogaland        ((-16941.25 6609897, -24653.9…
 5 3242          Hurdal      32           Akershus        ((264869.7 6717404, 264819.1 …
 6 5057          Ørland      50           Trøndelag       ((240772.2 7109470, 237510.9 …
 7 1539          Rauma       15           Møre og Romsdal ((129987.7 6972241, 128964 69…
 8 4637          Hyllestad   46           Vestland        ((-29129.98 6833953, -30155.6…
 9 4626          Øygarden    46           Vestland        ((-55999.97 6776596, -62120.5…
10 5043          Raarvihke   50           Trøndelag       ((460932.2 7224208, 460797.2 …
# ℹ 349 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000AdministrativeOmrader_GML.gml}
}
\usage{
Kommune
}
\description{
Kommune
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
