% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Kommunegrense}
\alias{Kommunegrense}
\title{Kommunegrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 779 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -92079.54 ymin: 6428945 xmax: 1111593 ymax: 7962288
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 779 × 4
   målemetode nøyaktighet oppdateringsdato                                        grense
 *      <int>       <int> <date>                                        <LINESTRING [m]>
 1         64       50000 2023-12-18                (212825.9 6975224, 215923.6 6977410)
 2         64       50000 2023-12-18       (384820 7447740, 376029.1 7443716, 354846.2 …
 3         64       50000 2023-12-18       (217522.9 7065976, 214808.9 7072861, 214709.…
 4         64       50000 2023-12-18       (283687.3 7047593, 281353.1 7047882, 278574.…
 5         64       50000 2023-12-18       (-50177.25 6592310, -56491.65 6588982, -7621…
 6         64       50000 2023-12-18       (249779.4 6634659, 250057.6 6636242, 251342.…
 7         64       50000 2023-12-18       (275672.4 7051059, 277376.8 7057321, 288025.…
 8         64       50000 2023-12-18              (-36231.25 6702712, -39434.61 6710972)
 9         64       50000 2023-12-18       (-38226.65 6690788, -42032.52 6688376, -5303…
10         64       50000 2023-12-18                (358383.6 7358348, 365421.9 7360393)
# ℹ 769 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000AdministrativeOmrader_GML.gml}
}
\usage{
Kommunegrense
}
\description{
Kommunegrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
