% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Kommunegrense}
\alias{Kommunegrense}
\title{Kommunegrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 779 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -92079.54 ymin: 6428945 xmax: 1111593 ymax: 7962288
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 779 × 4
   oppdateringsdato målemetode nøyaktighet                                        grense
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 2023-12-18       gen              50000 (-29064.66 6815856, -30515.85 6819011, -3299…
 2 2023-12-18       gen              50000 (407052.8 7146773, 409519 7137677, 413801.1 …
 3 2023-12-18       gen              50000 (280474 6556668, 279934.2 6557812, 280235.8 …
 4 2023-12-18       gen              50000 (153924.4 6824886, 156106.5 6823949, 157404.…
 5 2023-12-18       gen              50000 (228585.1 6666343, 230730.7 6672037, 232079.…
 6 2023-12-18       gen              50000 (916230.4 7677406, 915996.1 7677036, 915930.…
 7 2023-12-18       gen              50000 (291204.9 7054560, 293001.9 7053752, 294587.…
 8 2023-12-18       gen              50000 (229805 6816702, 229502.8 6816867, 227070.6 …
 9 2023-12-18       gen              50000 (337699.8 7055861, 336239 7053922, 335235.5 …
10 2023-12-18       gen              50000 (278254 6561019, 277129.8 6559350, 269541.4 …
# ℹ 769 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000AdministrativeOmrader_GML.gml}
}
\usage{
Kommunegrense
}
\description{
Kommunegrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
