% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Kystkontur}
\alias{Kystkontur}
\title{Kystkontur}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 6713 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -76208 ymin: 6450245 xmax: 1105197 ymax: 7939986
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 6,713 × 4
   oppdateringsdato målemetode nøyaktighet                                        grense
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 2017-01-05       gen              50000 (902487 7865963, 903053.3 7866784, 903238.3 …
 2 2017-01-05       gen              50000 (976706 7865961, 976057 7865890, 975937 7865…
 3 2017-01-05       gen              50000 (938859 7865346, 938634 7865431, 938600 7865…
 4 2017-01-05       gen              50000 (985334 7869249, 985141 7869102, 984923 7868…
 5 2017-01-05       gen              50000 (993619 7875506, 993685 7875740, 993824 7875…
 6 2017-01-05       gen              50000 (898637 7876251, 898820 7876462, 899327 7877…
 7 2017-01-05       gen              50000 (8e+05 7875160, 799887 7875292, 799910 78759…
 8 2017-01-05       gen              50000 (8e+05 7874601, 800081 7874648, 800212 78749…
 9 2017-01-05       gen              50000 (999704 7874550, 999947 7875014, 999905 7876…
10 2017-01-05       gen              50000 (8e+05 7874095, 799976 7874121, 799808 78744…
# ℹ 6,703 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
Kystkontur
}
\description{
Kystkontur
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
