% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Lufthavn}
\alias{Lufthavn}
\title{Lufthavn}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 75 features and 9 fields
Geometry type: POINT
Dimension:     XY
Bounding box:  xmin: -57805.33 ymin: 6472154 xmax: 1096256 ymax: 7929664
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 75 × 10
   oppdateringsdato målemetode nøyaktighet lufthavntype trafikktype iataKode icaoKode
 * <date>           <chr>            <int> <chr>        <chr>       <chr>    <chr>   
 1 2017-01-05       gen              50000 L            N           NTB      ENNO    
 2 2017-01-05       gen              50000 L            N           HFT      ENHF    
 3 2017-01-05       gen              50000 L            N           BJF      ENBS    
 4 2017-01-05       gen              50000 L            N           HVG      ENHV    
 5 2017-01-05       gen              50000 L            I           SVG      ENZV    
 6 2017-01-05       gen              50000 L            I           TRD      ENVA    
 7 2017-01-05       gen              50000 L            I           HAU      ENHD    
 8 2017-01-05       gen              50000 L            I           OSL      ENGM    
 9 2017-01-05       gen              50000 L            N           BVG      ENBV    
10 2017-01-05       gen              50000 L            I           BOO      ENBO    
# ℹ 65 more rows
# ℹ 3 more variables: lufthavneier <chr>, navn <chr>, posisjon <POINT [m]>
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
Lufthavn
}
\description{
Lufthavn
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
