% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Myr}
\alias{Myr}
\title{Myr}
\format{
\preformatted{Simple feature collection with 2864 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -58988.99 ymin: 6468671 xmax: 1114125 ymax: 7920789
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 2,864 x 2
   oppdateringsdato                                                            geometry
 * <date>                                                                 <POLYGON [m]>
 1 NA               ((1094181 7877371, 1093946 7877833, 1093606 7878083, 1093565 78781~
 2 NA               ((906559 7880343, 906520.8 7880385, 906258 7880671, 906366 7881104~
 3 NA               ((895923 7714082, 895948.3 7714045, 896047.4 7713899, 896107 77138~
 4 NA               ((916960 7878352, 916224.6 7877552, 915656 7877523, 915459.8 78774~
 5 NA               ((924175 7900000, 924165 7899640, 924454 7899152, 924518 7899044, ~
 6 NA               ((973162 7916613, 973042 7916825, 973099 7917901, 973123.5 7918016~
 7 NA               ((919098.8 7906979, 918368.8 7907024, 918298.8 7907347, 918457.8 7~
 8 NA               ((828119.5 7900000, 828103 7899976, 827787 7899808, 827540 7899788~
 9 NA               ((834861 7778528, 834958 7777820, 834775 7777612, 834551 7777515, ~
10 NA               ((890709.8 7896800, 890934 7896785, 891060 7896829, 890909.2 78964~
# ... with 2,854 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
Myr
}
\description{
Myr
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
