% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Myr}
\alias{Myr}
\title{Myr}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 2864 features and 0 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -58988.99 ymin: 6468671 xmax: 1114125 ymax: 7920789
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 2,864 × 1
                                                                                  område
 *                                                                         <POLYGON [m]>
 1 ((1094181 7877371, 1094744 7878028, 1095456 7877976, 1095539 7878224, 1095154 787872…
 2 ((906559 7880343, 906762 7880351, 906938 7880462, 907461.1 7881142, 907670 7881414, …
 3 ((895923 7714082, 895407 7714307, 894851 7714136, 894692.7 7714006, 894448 7713806, …
 4 ((916960 7878352, 917651 7879363, 917539 7879523, 917372 7879590, 915697 7878560, 91…
 5 ((924175 7900000, 924183 7900287, 923936 7900317, 923272 7900144, 923005.1 7900000, …
 6 ((973162 7916613, 973261.2 7916634, 973263 7916634, 973826 7916843, 974207 7917209, …
 7 ((919098.8 7906979, 919493.8 7907154, 919492.8 7907506, 919297.8 7908066, 918976.8 7…
 8 ((828119.5 7900000, 828229 7900159, 828262 7900348, 827895 7900651, 827765.8 7901070…
 9 ((834861 7778528, 835010 7778641, 835414 7778946, 835009 7779303, 835003.2 7779367, …
10 ((1088649 7875529, 1088903 7875532, 1089434 7875963, 1089570 7876187, 1089578 787634…
# ℹ 2,854 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
Myr
}
\description{
Myr
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
