% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Naturverngrense}
\alias{Naturverngrense}
\title{Naturverngrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 53 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 1183.26 ymin: 6472365 xmax: 1088054 ymax: 7893205
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 53 × 4
   målemetode nøyaktighet oppdateringsdato                                        grense
 *      <int>       <int> <date>                                        <LINESTRING [m]>
 1         64       50000 NA               (406356.7 7112941, 406338 7112954, 404544.4 …
 2         64       50000 NA               (397533.8 7146337, 398420.4 7146316, 399882.…
 3         64       50000 NA               (412402.3 7150661, 412515.5 7145783, 419327.…
 4         64       50000 NA               (334509.8 7006372, 330199.5 7004694, 328981.…
 5         64       50000 NA               (470434 7404044, 470471.4 7404021, 474319.2 …
 6         64       50000 NA               (863873 7658542, 863904.8 7658597, 864120.4 …
 7         64       50000 NA               (209157.5 6943996, 207715.7 6944281, 206807.…
 8         64       50000 NA               (101980.8 6893894, 102246.3 6895059, 102258.…
 9         64       50000 NA               (183631.6 6913086, 182381.1 6914390, 182037.…
10         64       50000 NA               (421883.6 7557498, 422657.2 7558012, 422702.…
# ℹ 43 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Restriksjonsomrader_GML.gml}
}
\usage{
Naturverngrense
}
\description{
Naturverngrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
