% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Naturverngrense}
\alias{Naturverngrense}
\title{Naturverngrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 54 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 1183.26 ymin: 6472365 xmax: 1088054 ymax: 7893205
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 54 × 4
   oppdateringsdato målemetode nøyaktighet                                        grense
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 NA               gen              50000 (279324.1 6643718, 279278.3 6643619, 279292.…
 2 NA               gen              50000 (406356.7 7112941, 406338 7112954, 404544.4 …
 3 NA               gen              50000 (397533.8 7146337, 398420.4 7146316, 399882.…
 4 NA               gen              50000 (863873 7658542, 863904.8 7658597, 864120.4 …
 5 NA               gen              50000 (209157.5 6943996, 207715.7 6944281, 206807.…
 6 NA               gen              50000 (412402.3 7150661, 412515.5 7145783, 419327.…
 7 NA               gen              50000 (470434 7404044, 470471.4 7404021, 474319.2 …
 8 NA               gen              50000 (183631.6 6913086, 182381.1 6914390, 182037.…
 9 NA               gen              50000 (421883.6 7557498, 422657.2 7558012, 422702.…
10 NA               gen              50000 (163398.7 6516399, 166482.2 6513639, 155821.…
# ℹ 44 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Restriksjonsomrader_GML.gml}
}
\usage{
Naturverngrense
}
\description{
Naturverngrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
