% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Naturvernomrade}
\alias{Naturvernomrade}
\title{Naturvernomrade}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 42 features and 4 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: 1183.26 ymin: 6472365 xmax: 1088054 ymax: 7893205
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 42 × 5
   oppdateringsdato navn           vernedato  verneform                           område
 * <date>           <chr>          <date>     <chr>                        <POLYGON [m]>
 1 2017-01-05       Sjunkhatten    2010-02-05 NP        ((494981.1 7472050, 495072.8 74…
 2 2017-01-05       Rohkunborri    2011-02-25 NP        ((642141.4 7608935, 645026.2 76…
 3 2017-01-05       Hallingskarvet 2006-12-22 NP        ((92235.33 6740841, 92255.15 67…
 4 2017-01-05       Folgefonna     2005-04-29 NP        ((28219.27 6691623, 28162.51 66…
 5 2017-01-05       Ånderdalen     2004-06-04 NP        ((585497 7677287, 586045 767750…
 6 2017-01-05       Jotunheimen    1980-12-05 NP        ((134939 6826311, 136589.2 6826…
 7 2017-01-05       Ytre Hvaler    2009-06-26 NP        ((271280.2 6550636, 271187.1 65…
 8 2017-01-05       Øvre Pasvik    2003-08-29 NP        ((1043856 7739675, 1044377 7737…
 9 2017-01-05       Stabbursdalen  2002-12-20 NP        ((845960.9 7801580, 846028.5 77…
10 2017-01-05       Reisa          1986-11-28 NP        ((753705.9 7706934, 756451.2 77…
# ℹ 32 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Restriksjonsomrader_GML.gml}
}
\usage{
Naturvernomrade
}
\description{
Naturvernområde
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
