% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Naturvernomrade}
\alias{Naturvernomrade}
\title{Naturvernomrade}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 43 features and 3 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: 1183.26 ymin: 6472365 xmax: 1088054 ymax: 7893205
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 43 × 4
   vernedato  verneform navn                                                      område
 * <date>     <chr>     <chr>                                              <POLYGON [m]>
 1 2010-02-05 NP        Sjunkhatten    ((494981.1 7472050, 495072.8 7472161, 495640.3 7…
 2 2011-02-25 NP        Rohkunborri    ((642141.4 7608935, 645026.2 7606453, 647994.3 7…
 3 2006-12-22 NP        Hallingskarvet ((92235.33 6740841, 92255.15 6740824, 92680.45 6…
 4 2005-04-29 NP        Folgefonna     ((28219.27 6691623, 28162.51 6692038, 28225.63 6…
 5 2004-06-04 NP        Ånderdalen     ((585497 7677287, 586045 7677500, 586416 7676993…
 6 1980-12-05 NP        Jotunheimen    ((134939 6826311, 136589.2 6826480, 138002 68271…
 7 2009-06-26 NP        Ytre Hvaler    ((271280.2 6550636, 271187.1 6550720, 271131.6 6…
 8 2003-08-29 NP        Øvre Pasvik    ((1043856 7739675, 1044377 7737311, 1044812 7735…
 9 2002-12-20 NP        Stabbursdalen  ((845960.9 7801580, 846028.5 7797680, 848842.2 7…
10 1986-11-28 NP        Reisa          ((753705.9 7706934, 756451.2 7704942, 757967.3 7…
# ℹ 33 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Restriksjonsomrader_GML.gml}
}
\usage{
Naturvernomrade
}
\description{
Naturvernområde
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
