% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Navigasjonsinstallasjon}
\alias{Navigasjonsinstallasjon}
\title{Navigasjonsinstallasjon}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 106 features and 3 fields
Geometry type: POINT
Dimension:     XY
Bounding box:  xmin: -75270 ymin: 6449852 xmax: 1098990 ymax: 7939496
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 106 × 4
   oppdateringsdato målemetode nøyaktighet           posisjon
 * <date>           <chr>            <int>        <POINT [m]>
 1 2017-01-05       gen              50000   (253016 6639164)
 2 2017-01-05       gen              50000 (248095.2 6585792)
 3 2017-01-05       gen              50000 (250038.4 6565999)
 4 2017-01-05       gen              50000   (261511 6564372)
 5 2017-01-05       gen              50000   (243201 6551650)
 6 2017-01-05       gen              50000   (198903 6550952)
 7 2017-01-05       gen              50000   (271787 6548674)
 8 2017-01-05       gen              50000   (257732 6547330)
 9 2017-01-05       gen              50000   (221145 6546755)
10 2017-01-05       gen              50000   (208731 6544222)
# ℹ 96 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000BygningerOgAnlegg_GML.gml}
}
\usage{
Navigasjonsinstallasjon
}
\description{
Navigasjonsinstallasjon
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
