% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Passasjerferjestrekning}
\alias{Passasjerferjestrekning}
\title{Passasjerferjestrekning}
\format{
\preformatted{Simple feature collection with 25 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 59772 ymin: 6544205 xmax: 445470 ymax: 7214682
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 25 x 4
   målemetode nøyaktighet oppdateringsdato                                     geometry
 *      <int>       <int> <date>                                       <LINESTRING [m]>
 1         64       50000 2007-10-24       (121388.1 6691851, 121388.9 6691848, 121778~
 2         64       50000 2007-10-24       (104307.6 6711695, 104313.5 6711700, 104440~
 3         64       50000 2004-03-01       (343821.1 6917814, 343819.5 6917813, 343641~
 4         64       50000 2004-03-01       (337579.3 6903838, 337589.7 6903833, 338321~
 5         64       50000 2012-03-28       (125379 6649884, 125411.2 6649978, 125414.9~
 6         64       50000 2007-10-24       (339120.1 6924195, 339107.7 6924078, 339136~
 7         64       50000 2007-10-24       (170571.7 6833873, 168819.4 6833780, 167963~
 8         64       50000 2007-10-24       (161164.3 6833755, 160886.8 6833311, 159291~
 9         64       50000 2004-03-01       (340267.1 6912784, 340279 6912775, 340560.9~
10         64       50000 2018-04-11       (293104.8 6693827, 293168.3 6694144, 293137~
# ... with 15 more rows
}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Samferdsel_GML.gml}
}
\usage{
Passasjerferjestrekning
}
\description{
Passasjerferjestrekning
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
