% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Riksgrense}
\alias{Riksgrense}
\title{Riksgrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 40 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: 245176.8 ymin: 6521821 xmax: 1121941 ymax: 7867303
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 40 × 4
   oppdateringsdato målemetode nøyaktighet                                        grense
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 2023-12-18       gen              50000 (314938.8 6614098, 315066.2 6614272, 318070.…
 2 2023-12-18       gen              50000 (311244.9 6548450, 311784.7 6548921, 312205.…
 3 2023-12-18       gen              50000 (1111593 7849213, 1109098 7844651, 1105118 7…
 4 2023-12-18       gen              50000 (279596.3 6555228, 283734.6 6555924, 284985.…
 5 2023-12-18       gen              50000 (412808.6 7108866, 417042.6 7107857, 420919.…
 6 2023-12-18       gen              50000 (353736.6 6960754, 353016.3 6966019, 352483.…
 7 2023-12-18       gen              50000 (523677.4 7382194, 524664.7 7383241, 527543.…
 8 2023-12-18       gen              50000 (319002.8 6577574, 318988.2 6578328, 319087.…
 9 2023-12-18       gen              50000 (738347 7692234, 740578.4 7695439, 744627.3 …
10 2023-12-18       gen              50000 (352618.2 6652316, 353733.1 6654174, 354829.…
# ℹ 30 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000AdministrativeOmrader_GML.gml}
}
\usage{
Riksgrense
}
\description{
Riksgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
