% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Skog}
\alias{Skog}
\title{Skog}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 2884 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -56133 ymin: 6450729 xmax: 1114945 ymax: 7921538
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 2,884 × 2
   oppdateringsdato                                                               område
 * <date>                                                                  <POLYGON [m]>
 1 2017-01-05       ((951956 7911719, 952745 7911759, 953274 7911902, 953446 7912069, 9…
 2 2017-01-05       ((953052 7910658, 953046 7910361, 953195 7910215, 953414 7910162, 9…
 3 2017-01-05       ((956097 7907628, 956631 7907668, 956812 7907732, 957150 7907851, 9…
 4 2017-01-05       ((951364.2 7900000, 951467 7899954, 951700.1 7899985, 951732 789998…
 5 2017-01-05       ((1e+06 7873838, 1e+06 7874608, 999704 7874550, 999521 7874200, 999…
 6 2017-01-05       ((977579 7907649, 977780.4 7907798, 977827.6 7907833, 977921 790790…
 7 2017-01-05       ((951742.7 7900000, 951792 7900051, 951879 7900140, 951990 7900716,…
 8 2017-01-05       ((908181 7869131, 908363 7869230, 908393 7869616, 908224 7869896, 9…
 9 2017-01-05       ((852935.1 7890473, 852922 7890430, 852887 7890312, 852766 7890352,…
10 2017-01-05       ((913655 7876619, 913530.2 7876864, 913505 7876913, 912854 7876842,…
# ℹ 2,874 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
Skog
}
\description{
Skog
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
