% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Skog}
\alias{Skog}
\title{Skog}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 2884 features and 0 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -56133 ymin: 6450729 xmax: 1114945 ymax: 7921538
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 2,884 × 1
                                                                                  område
 *                                                                         <POLYGON [m]>
 1 ((951956 7911719, 952745 7911759, 953274 7911902, 953446 7912069, 953475.7 7912151, …
 2 ((953052 7910658, 953046 7910361, 953195 7910215, 953414 7910162, 953841 7910607, 95…
 3 ((956097 7907628, 956631 7907668, 956812 7907732, 957150 7907851, 957150.4 7907852, …
 4 ((951364.2 7900000, 951467 7899954, 951700.1 7899985, 951732 7899989, 951742.7 79000…
 5 ((1e+06 7873838, 1e+06 7874608, 999704 7874550, 999521 7874200, 999617 7874038, 9998…
 6 ((977579 7907649, 977780.4 7907798, 977827.6 7907833, 977921 7907902, 977913 7908045…
 7 ((951742.7 7900000, 951792 7900051, 951879 7900140, 951990 7900716, 951882.1 7900899…
 8 ((908181 7869131, 908363 7869230, 908393 7869616, 908224 7869896, 907738 7870104, 90…
 9 ((852935.1 7890473, 852922 7890430, 852887 7890312, 852766 7890352, 852710 7890613, …
10 ((913655 7876619, 913530.2 7876864, 913505 7876913, 912854 7876842, 912457 7876682, …
# ℹ 2,874 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
Skog
}
\description{
Skog
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
