% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{SnoIsbre}
\alias{SnoIsbre}
\title{SnoIsbre}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 118 features and 1 field
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -5664 ymin: 6649920 xmax: 811950.2 ymax: 7837518
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 118 × 2
   oppdateringsdato                                                               område
 * <date>                                                                  <POLYGON [m]>
 1 2017-01-05       ((763484.8 7808343, 763449.6 7808500, 763383 7808796, 763318 780908…
 2 2017-01-05       ((768702.6 7802638, 768646 7802682, 767764.1 7802799, 767756 780280…
 3 2017-01-05       ((691542.8 7706433, 691562 7706747, 691754.3 7706959, 691840.1 7707…
 4 2017-01-05       ((477654.5 7401440, 477709.8 7402206, 477704.6 7402265, 477637.6 74…
 5 2017-01-05       ((452352 7413498, 453509.7 7413786, 453612.5 7413812, 453918 741388…
 6 2017-01-05       ((669816 7636044, 669476.6 7636487, 669014.7 7637090, 668907.1 7637…
 7 2017-01-05       ((474806.6 7397547, 475228.2 7397896, 475231 7397909, 475249.6 7397…
 8 2017-01-05       ((471712.7 7432991, 472026.5 7432991, 472220.3 7433085, 472371.6 74…
 9 2017-01-05       ((591095.2 7572683, 590885 7573071, 590410.2 7573158, 590183.3 7573…
10 2017-01-05       ((113807.8 6938262, 113896 6938511, 113934.4 6938740, 113972 693896…
# ℹ 108 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
SnoIsbre
}
\description{
SnøIsbre
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
