% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{SnoIsbre}
\alias{SnoIsbre}
\title{SnoIsbre}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 118 features and 0 fields
Geometry type: POLYGON
Dimension:     XY
Bounding box:  xmin: -5664 ymin: 6649920 xmax: 811950.2 ymax: 7837518
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 118 × 1
                                                                                  område
 *                                                                         <POLYGON [m]>
 1 ((763484.8 7808343, 763449.6 7808500, 763383 7808796, 763318 7809085, 763195 7809223…
 2 ((768702.6 7802638, 768646 7802682, 767764.1 7802799, 767756 7802800, 767579.4 78030…
 3 ((691542.8 7706433, 691562 7706747, 691754.3 7706959, 691840.1 7707053, 692056.7 770…
 4 ((477654.5 7401440, 477709.8 7402206, 477704.6 7402265, 477637.6 7403028, 477614.4 7…
 5 ((452352 7413498, 453509.7 7413786, 453612.5 7413812, 453918 7413888, 454576.5 74141…
 6 ((669816 7636044, 669476.6 7636487, 669014.7 7637090, 668907.1 7637231, 669074.8 763…
 7 ((474806.6 7397547, 475228.2 7397896, 475231 7397909, 475249.6 7397991, 475470.4 739…
 8 ((471712.7 7432991, 472026.5 7432991, 472220.3 7433085, 472371.6 7433158, 472293.7 7…
 9 ((591095.2 7572683, 590885 7573071, 590410.2 7573158, 590183.3 7573200, 589881 75732…
10 ((113807.8 6938262, 113896 6938511, 113934.4 6938740, 113972 6938964, 113969.1 69390…
# ℹ 108 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Arealdekke_GML.gml}
}
\usage{
SnoIsbre
}
\description{
SnøIsbre
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
