% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Stasjon}
\alias{Stasjon}
\title{Stasjon}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 176 features and 4 fields
Geometry type: POINT
Dimension:     XY
Bounding box:  xmin: -42220.15 ymin: 6466305 xmax: 625842.6 ymax: 7594511
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 176 × 5
   oppdateringsdato målemetode nøyaktighet navn                       posisjon
 * <date>           <chr>            <int> <chr>                   <POINT [m]>
 1 2017-01-05       gen              50000 Varhaug         (-42220.15 6535489)
 2 2017-01-05       gen              50000 Gjerstad         (155611.2 6540993)
 3 2017-01-05       gen              50000 Bryne           (-40262.43 6548171)
 4 2017-01-05       gen              50000 Larvik           (215038.1 6556235)
 5 2017-01-05       gen              50000 Halden           (292989.7 6559010)
 6 2017-01-05       gen              50000 Sandnes sentrum (-33323.07 6560631)
 7 2017-01-05       gen              50000 Jåttåvågen      (-32796.72 6567525)
 8 2017-01-05       gen              50000 Drangedal        (160314.6 6565885)
 9 2017-01-05       gen              50000 Porsgrunn        (194724.8 6568049)
10 2017-01-05       gen              50000 Fredrikstad      (268871.8 6570367)
# ℹ 166 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Samferdsel_GML.gml}
}
\usage{
Stasjon
}
\description{
Stasjon
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
