% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{StedsnavnTekst}
\alias{StedsnavnTekst}
\title{StedsnavnTekst}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 5520 features and 17 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -76547.7 ymin: 6447274 xmax: 1114130 ymax: 7941194
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 5,520 × 18
   streng  oppdateringsdato datafangstdato tekstReferansePunktN…¹ tekstReferansePunktØst
 * <chr>   <date>           <date>                          <int>                  <int>
 1 Søre O… 2021-11-11       2017-01-06                          0                      0
 2 Strand  2021-11-11       2017-01-06                          0                      2
 3 Farest… 2021-11-11       2017-01-06                          0                      0
 4 Undheim 2021-11-11       2017-01-06                          0                      0
 5 Kjerri… 2021-11-11       2017-01-06                          0                      2
 6 Leines  2021-11-11       2017-01-06                          0                      1
 7 Tårnvi… 2021-11-11       2017-01-06                          0                      1
 8 Sandst… 2021-11-11       2020-09-25                          0                      0
 9 Grønni… 2021-11-11       2017-01-06                          0                      1
10 Sørli   2021-11-11       2017-01-06                          0                      0
# ℹ 5,510 more rows
# ℹ abbreviated name: ¹​tekstReferansePunktNord
# ℹ 13 more variables: tegnavstand <int>, skriftkode <int>, skrifttype <chr>,
#   referansemålestokk <int>, fulltekst <chr>, navneobjekthovedgruppe <int>,
#   navneobjektgruppe <int>, navneobjekttype <int>, språk <chr>,
#   stedsnavnsnummer <int>, skrivemåtenummer <int>, stedsnummer <int>,
#   geometri <LINESTRING [m]>
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Stedsnavn_GML.gml}
}
\usage{
StedsnavnTekst
}
\description{
StedsnavnTekst
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
