% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{StedsnavnTekst}
\alias{StedsnavnTekst}
\title{StedsnavnTekst}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 5526 features and 16 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -76547.7 ymin: 6447274 xmax: 1114130 ymax: 7941194
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 5,526 × 17
   streng     oppdateringsdato tekstReferansePunktN…¹ tekstReferansePunktØst tegnavstand
 * <chr>      <date>                            <int>                  <int>       <int>
 1 Søre Osen  2021-11-11                            0                      0           0
 2 Strand     2021-11-11                            0                      2           0
 3 Farestad   2021-11-11                            0                      0           0
 4 Undheim    2021-11-11                            0                      0           0
 5 Kjerringøy 2021-11-11                            0                      2           0
 6 Leines     2021-11-11                            0                      1           0
 7 Tårnvika   2021-11-11                            0                      1           0
 8 Sandstrand 2021-11-11                            0                      0           0
 9 Grønningf… 2021-11-11                            0                      1           0
10 Sørli      2021-11-11                            0                      0           0
# ℹ 5,516 more rows
# ℹ abbreviated name: ¹​tekstReferansePunktNord
# ℹ 12 more variables: skriftkode <int>, skrifttype <chr>, referansemålestokk <int>,
#   fulltekst <chr>, navneobjekthovedgruppe <chr>, navneobjektgruppe <chr>,
#   navneobjekttype <chr>, språk <chr>, stedsnavnnummer <int>, stedsnummer <int>,
#   skrivemåtenummer <int>, geometri <LINESTRING [m]>
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Stedsnavn_GML.gml}
}
\usage{
StedsnavnTekst
}
\description{
StedsnavnTekst
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
