% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Sti}
\alias{Sti}
\title{Sti}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 924 features and 4 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -39230 ymin: 6527264 xmax: 1022648 ymax: 7899446
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 924 × 5
   målemetode nøyaktighet oppdateringsdato merking                           senterlinje
 *      <int>       <int> <date>           <chr>                        <LINESTRING [m]>
 1         64       50000 2017-01-05       JA      (760049 7715757, 761258.9 7715493, 7…
 2         64       50000 2017-01-05       JA      (755209 7718900, 755143 7718899, 754…
 3         64       50000 2017-01-05       JA      (775455 7704871, 775480.9 7704839, 7…
 4         64       50000 2017-01-05       JA      (810550 7682223, 810782.5 7682333, 8…
 5         64       50000 2017-01-05       JA      (811940 7681316, 811899.4 7681203, 8…
 6         64       50000 2017-01-05       JA      (703898 7665254, 703428 7663204, 703…
 7         64       50000 2017-01-05       JA      (705539 7642898, 706005.6 7642486, 7…
 8         64       50000 2017-01-05       JA      (184989.5 7008291, 184963.5 7008383,…
 9         64       50000 2017-01-05       JA      (515892 7353961, 515794 7354181, 515…
10         64       50000 2017-01-05       JA      (147806.9 6921673, 147742 6922620, 1…
# ℹ 914 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Samferdsel_GML.gml}
}
\usage{
Sti
}
\description{
Sti
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
