% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Terrengpunkt}
\alias{Terrengpunkt}
\title{Terrengpunkt}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 553 features and 5 fields
Geometry type: POINT
Dimension:     XY
Bounding box:  xmin: -50222 ymin: 6464874 xmax: 1105697 ymax: 7936618
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 553 × 6
   målemetode nøyaktighet medium oppdateringsdato høyde         posisjon
 *      <int>       <int> <chr>  <date>           <int>      <POINT [m]>
 1         64       50000 T      2022-10-03        2258 (169248 6845853)
 2         64       50000 T      2022-10-03        2178 (229769 6875425)
 3         64       50000 T      2022-10-03        2014 (140460 6901277)
 4         64       50000 T      2022-10-03        1999 (120629 6919539)
 5         64       50000 T      2022-10-03        1964 (149265 6936766)
 6         64       50000 T      2022-10-03        1951 (179420 6924554)
 7         64       50000 T      2022-10-03        1933 (105371 6740759)
 8         64       50000 T      2022-10-03        1920 (134074 6787783)
 9         64       50000 T      2022-10-03        1916 (469997 7321657)
10         64       50000 T      2022-10-03        1908 (559686 7448403)
# ℹ 543 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Hoyde_GML.gml}
}
\usage{
Terrengpunkt
}
\description{
Terrengpunkt
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
