% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Territorialgrense}
\alias{Territorialgrense}
\title{Territorialgrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 78 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -99551.21 ymin: 6426048 xmax: 1121941 ymax: 7962744
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 78 × 4
   målemetode nøyaktighet oppdateringsdato                                        grense
 *      <int>       <int> <date>                                        <LINESTRING [m]>
 1         64       50000 2023-12-18              (-79642.38 6687459, -79575.13 6709024)
 2         64       50000 2023-12-18       (-86165.9 6792747, -87860.58 6809231, -87940…
 3         64       50000 2023-12-18       (394959.1 7562014, 400056.4 7571103, 400564.…
 4         64       50000 2023-12-18       (927892.2 7959494, 951360.7 7962151, 952358.…
 5         64       50000 2023-12-18       (606182.7 7787101, 609670.6 7792678, 618719.…
 6         64       50000 2023-12-18       (-20575.89 6963321, -19545.18 6964471, -1885…
 7         64       50000 2023-12-18       (11955.23 6438285, -1695.79 6445937, -2346.4…
 8         64       50000 2023-12-18       (21146.36 6433669, 20724.18 6433811, 19781.6…
 9         64       50000 2023-12-18       (-66423.89 6897184, -51894.04 6920807, -5070…
10         64       50000 2023-12-18                (362047.3 7414276, 354846.2 7437006)
# ℹ 68 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000AdministrativeOmrader_GML.gml}
}
\usage{
Territorialgrense
}
\description{
Territorialgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
