% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{Territorialgrense}
\alias{Territorialgrense}
\title{Territorialgrense}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 78 features and 3 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -99551.21 ymin: 6426048 xmax: 1121941 ymax: 7962744
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 78 × 4
   oppdateringsdato målemetode nøyaktighet                                        grense
 * <date>           <chr>            <int>                              <LINESTRING [m]>
 1 2023-12-18       gen              50000 (927892.2 7959494, 951360.7 7962151, 952358.…
 2 2023-12-18       gen              50000 (606182.7 7787101, 609670.6 7792678, 618719.…
 3 2023-12-18       gen              50000 (-20575.89 6963321, -19545.18 6964471, -1885…
 4 2023-12-18       gen              50000 (11955.23 6438285, -1695.79 6445937, -2346.4…
 5 2023-12-18       gen              50000 (21146.36 6433669, 20724.18 6433811, 19781.6…
 6 2023-12-18       gen              50000 (-66423.89 6897184, -51894.04 6920807, -5070…
 7 2023-12-18       gen              50000          (362047.3 7414276, 354846.2 7437006)
 8 2023-12-18       gen              50000 (380921.3 7531107, 386002.7 7545389, 386358.…
 9 2023-12-18       gen              50000          (298752.7 7275338, 300545.5 7279109)
10 2023-12-18       gen              50000         (-84440.08 6775960, -86165.9 6792747)
# ℹ 68 more rows
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000AdministrativeOmrader_GML.gml}
}
\usage{
Territorialgrense
}
\description{
Territorialgrense
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
