% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadoc.R
\docType{data}
\name{VegSenterlinje}
\alias{VegSenterlinje}
\title{VegSenterlinje}
\format{
\if{html}{\out{<div class="sourceCode">}}\preformatted{Simple feature collection with 13200 features and 8 fields
Geometry type: LINESTRING
Dimension:     XY
Bounding box:  xmin: -74500 ymin: 6452103 xmax: 1112920 ymax: 7939337
Projected CRS: ETRS89 / UTM zone 33N
# A tibble: 13,200 × 9
   målemetode nøyaktighet medium oppdateringsdato vegkategori vegstatus vegnummer
 *      <int>       <int> <chr>  <date>           <chr>       <chr>         <int>
 1         64       50000 T      2017-01-05       K           V                NA
 2         64       50000 T      2017-01-05       P           V                NA
 3         64       50000 T      2017-01-05       K           V                NA
 4         64       50000 T      2017-01-05       K           V                NA
 5         64       50000 T      2017-01-05       P           V                NA
 6         64       50000 T      2017-01-05       K           V                NA
 7         64       50000 T      2017-01-05       P           V                NA
 8         64       50000 T      2017-01-05       P           V                NA
 9         64       50000 T      2017-01-05       K           V                NA
10         64       50000 T      2017-01-05       K           V                NA
# ℹ 13,190 more rows
# ℹ 2 more variables: motorvegtype <chr>, senterlinje <LINESTRING [m]>
# ℹ Use `print(n = ...)` to see more rows
}\if{html}{\out{</div>}}
}
\source{
\code{Basisdata_0000_Norge_25833_N1000Samferdsel_GML.gml}
}
\usage{
VegSenterlinje
}
\description{
VegSenterlinje
}
\author{
© \href{https://kartverket.no/}{Kartverket}
}
\keyword{datasets}
